refactor(cli): extract gc and cache modules from main.rs (#339)
Move the GC orchestrator, auto-GC tiers, and cache/status/version
commands out of main.rs into dedicated sibling modules. Tests still
live in src/tests.rs from the prior chunk.

Refs #339
