+++ crates/soldr-cli/src/main.rs
@@ -4,10 +4,9 @@ use sha2::{Digest, Sha256};
 use soldr_core::{suppress_windows_console_window, SoldrError, SoldrPaths};
 use soldr_fetch::VersionSpec;
 use std::collections::BTreeSet;
-use std::io::Write;
-use std::sync::{mpsc, Arc, Mutex};
-use std::time::Duration;
 
+mod cache;
+mod gc;
 mod linker;
 mod self_relocate;
 
@@ -19,8 +18,8 @@ const TEST_FREE_DISK_BYTES_ENV_VAR: &str = "SOLDR_TEST_FREE_DISK_BYTES";
 /// Overrides the default `nightly` toolchain used by `soldr gc cargo`
 /// to invoke cargo's unstable `-Zgc clean gc`. The `--toolchain` flag
 /// on `gc cargo` / `gc sweep` always takes precedence.
-const SOLDR_GC_CARGO_TOOLCHAIN_ENV_VAR: &str = "SOLDR_GC_CARGO_TOOLCHAIN";
-const JSON_SCHEMA_VERSION: u32 = 1;
+pub(crate) const SOLDR_GC_CARGO_TOOLCHAIN_ENV_VAR: &str = "SOLDR_GC_CARGO_TOOLCHAIN";
+pub(crate) const JSON_SCHEMA_VERSION: u32 = 1;
 const LOW_DISK_WARNING_THRESHOLD_BYTES: u64 = 2 * 1024 * 1024 * 1024;
 const RUSTC_WRAPPER_OVERRIDE_ENV_VAR: &str = "SOLDR_RUSTC_WRAPPER";
 const CARGO_PROFILE_DEV_DEBUG_ENV_VAR: &str = "CARGO_PROFILE_DEV_DEBUG";
@@ -272,76 +271,76 @@ enum GcSubcommand {
 }
 
 #[derive(clap::Args)]
-struct GcCargoArgs {
+pub(crate) struct GcCargoArgs {
     /// Report the plan and exit without invoking cargo.
     #[arg(long)]
-    dry_run: bool,
+    pub(crate) dry_run: bool,
     /// Override the nightly toolchain. Defaults to
     /// `$SOLDR_GC_CARGO_TOOLCHAIN` if set, else `nightly`.
     #[arg(long, value_name = "TOOLCHAIN")]
-    toolchain: Option<String>,
+    pub(crate) toolchain: Option<String>,
     /// Forwarded directly to cargo `--max-src-age`.
     #[arg(long, value_name = "DURATION")]
-    max_src_age: Option<String>,
+    pub(crate) max_src_age: Option<String>,
     /// Forwarded directly to cargo `--max-crate-age`.
     #[arg(long, value_name = "DURATION")]
-    max_crate_age: Option<String>,
+    pub(crate) max_crate_age: Option<String>,
     /// Forwarded directly to cargo `--max-index-age`.
     #[arg(long, value_name = "DURATION")]
-    max_index_age: Option<String>,
+    pub(crate) max_index_age: Option<String>,
     /// Forwarded directly to cargo `--max-git-co-age`.
     #[arg(long, value_name = "DURATION")]
-    max_git_co_age: Option<String>,
+    pub(crate) max_git_co_age: Option<String>,
     /// Forwarded directly to cargo `--max-git-db-age`.
     #[arg(long, value_name = "DURATION")]
-    max_git_db_age: Option<String>,
+    pub(crate) max_git_db_age: Option<String>,
     /// Forwarded directly to cargo `--max-download-age`.
     #[arg(long, value_name = "DURATION")]
-    max_download_age: Option<String>,
+    pub(crate) max_download_age: Option<String>,
     /// Forwarded directly to cargo `--max-src-size`.
     #[arg(long, value_name = "SIZE")]
-    max_src_size: Option<String>,
+    pub(crate) max_src_size: Option<String>,
     /// Forwarded directly to cargo `--max-crate-size`.
     #[arg(long, value_name = "SIZE")]
-    max_crate_size: Option<String>,
+    pub(crate) max_crate_size: Option<String>,
     /// Forwarded directly to cargo `--max-git-size`.
     #[arg(long, value_name = "SIZE")]
-    max_git_size: Option<String>,
+    pub(crate) max_git_size: Option<String>,
     /// Forwarded directly to cargo `--max-download-size`.
     #[arg(long, value_name = "SIZE")]
-    max_download_size: Option<String>,
+    pub(crate) max_download_size: Option<String>,
     /// Emit the stable machine-facing JSON form for this command.
     #[arg(long)]
-    json: bool,
+    pub(crate) json: bool,
 }
 
 #[derive(clap::Args)]
-struct GcSweepArgs {
+pub(crate) struct GcSweepArgs {
     /// Delete every eligible target/ candidate without prompting (used
     /// when the orchestrator runs the soldr target purge stage).
     #[arg(long)]
-    all: bool,
+    pub(crate) all: bool,
     /// Plan and report without deleting anything.
     #[arg(long)]
-    dry_run: bool,
+    pub(crate) dry_run: bool,
     /// Run cargo's `clean gc`. Default is on; pass `--no-cargo-gc` to
     /// skip cargo entirely. `--cargo-gc` is accepted but is the
     /// default.
     #[arg(long, conflicts_with = "no_cargo_gc")]
-    cargo_gc: bool,
+    pub(crate) cargo_gc: bool,
     /// Skip cargo's `clean gc` (e.g. on CI runners with no nightly).
     #[arg(long, conflicts_with = "cargo_gc")]
-    no_cargo_gc: bool,
+    pub(crate) no_cargo_gc: bool,
     /// After the standard pipeline, run cargo's `clean gc` again with
     /// tighter ages
     /// (`--max-src-age=7days --max-crate-age=14days --max-git-co-age=7days`).
     /// Floor: each value is clamped to `auto_gc.min_age_secs` before
     /// being forwarded.
     #[arg(long)]
-    aggressive: bool,
+    pub(crate) aggressive: bool,
     /// Emit the stable machine-facing JSON form for this command.
     #[arg(long)]
-    json: bool,
+    pub(crate) json: bool,
 }
 
 #[derive(Subcommand)]
@@ -505,25 +504,25 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
             std::process::exit(run_doctor(json)?);
         }
         Commands::Status { json } => {
-            let output = collect_status_output(cache_enabled)?;
+            let output = cache::collect_status_output(cache_enabled)?;
             if json {
-                print_json(&output)?;
+                cache::print_json(&output)?;
             } else {
-                print_status_output(&output);
+                cache::print_status_output(&output);
             }
         }
         Commands::Clean => {
-            clear_zccache_cache()?;
+            cache::clear_zccache_cache()?;
         }
         Commands::Purge => {
-            purge_soldr_cache()?;
+            cache::purge_soldr_cache()?;
         }
         Commands::Config => {
             println!("(config not yet implemented)");
         }
         Commands::Cache { json, command } => match command {
             Some(CacheSubcommand::Report { json: report_json }) => {
-                run_cache_report_command(report_json || json)?;
+                cache::run_cache_report_command(report_json || json)?;
             }
             Some(CacheSubcommand::PruneTarget {
                 path,
@@ -536,21 +535,25 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
                 // Either flag pair maps onto the same boolean; `dry_run`
                 // is the documented default so we accept it explicitly.
                 let _ = dry_run;
-                run_cache_prune_target_command(path, effective_dry_run, prune_json || json)?;
+                cache::run_cache_prune_target_command(
+                    path,
+                    effective_dry_run,
+                    prune_json || json,
+                )?;
             }
             None => {
-                let output = collect_cache_output()?;
+                let output = cache::collect_cache_output()?;
                 if json {
-                    print_json(&output)?;
+                    cache::print_json(&output)?;
                 } else {
-                    print_cache_output(&output);
+                    cache::print_cache_output(&output);
                 }
             }
         },
         Commands::Version { json } => {
-            let output = version_output();
+            let output = cache::version_output();
             if json {
-                print_json(&output)?;
+                cache::print_json(&output)?;
             } else {
                 println!("soldr {}", output.soldr_version);
             }
@@ -580,36 +583,36 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
                     older_than,
                     larger_than,
                     json,
-                }) => GcInvocation {
-                    mode: GcMode::Purge { all },
+                }) => gc::GcInvocation {
+                    mode: gc::GcMode::Purge { all },
                     older_than,
                     larger_than,
                     json,
                 },
                 Some(GcSubcommand::List { json }) => {
-                    run_gc_list_command(json)?;
+                    gc::run_gc_list_command(json)?;
                     return Ok(());
                 }
                 Some(GcSubcommand::Cargo(args)) => {
-                    run_gc_cargo_command(*args)?;
+                    gc::run_gc_cargo_command(*args)?;
                     return Ok(());
                 }
                 Some(GcSubcommand::Locations { json }) => {
-                    run_gc_locations_command(json)?;
+                    gc::run_gc_locations_command(json)?;
                     return Ok(());
                 }
                 Some(GcSubcommand::Sweep(args)) => {
-                    run_gc_sweep_command(*args)?;
+                    gc::run_gc_sweep_command(*args)?;
                     return Ok(());
                 }
-                None => GcInvocation {
-                    mode: GcMode::Summary,
+                None => gc::GcInvocation {
+                    mode: gc::GcMode::Summary,
                     older_than,
                     larger_than,
                     json,
                 },
             };
-            run_gc_command(invocation)?;
+            gc::run_gc_command(invocation)?;
         }
         Commands::External(args) => {
             if args.is_empty() {
@@ -1208,7 +1211,7 @@ fn run_doctor(json: bool) -> Result<i32, SoldrError> {
                 missing_components: Vec::new(),
                 missing_targets: Vec::new(),
             };
-            print_json(&output)?;
+            cache::print_json(&output)?;
         } else if manifest_present {
             println!(
                 "manifest: {} (present but no [toolchain] channel declared)",
@@ -1285,7 +1288,7 @@ fn run_doctor(json: bool) -> Result<i32, SoldrError> {
             missing_components,
             missing_targets,
         };
-        print_json(&output)?;
+        cache::print_json(&output)?;
     } else {
         print_doctor_human(
             &manifest_path,
@@ -1719,10 +1722,10 @@ async fn run_cargo_front_door(args: &[String], cache_enabled: bool) -> Result<i3
     if build_like_cargo {
         // Cargo front door only: keep startup/low-disk warnings off unrelated
         // commands and out of the rustc-wrapper hot path.
-        emit_startup_target_warning_if_due();
+        gc::emit_startup_target_warning_if_due();
         // Best-effort auto-GC trigger (issue #323). Runs on a detached
         // background thread; never blocks the build.
-        maybe_kick_auto_gc(&paths);
+        gc::maybe_kick_auto_gc(&paths);
     }
     let mut path_dirs: Vec<std::path::PathBuf> = Vec::with_capacity(1 + extra_bin_dirs.len());
     path_dirs.push(cargo_bin_dir);
@@ -3602,7 +3605,7 @@ fn stderr_should_use_color() -> bool {
     std::env::var_os("NO_COLOR").is_none() && std::io::stderr().is_terminal()
 }
 
-fn available_space(path: &std::path::Path) -> std::io::Result<u64> {
+pub(crate) fn available_space(path: &std::path::Path) -> std::io::Result<u64> {
     if let Some(raw) = std::env::var_os(TEST_FREE_DISK_BYTES_ENV_VAR) {
         let raw = raw.to_string_lossy();
         if raw.eq_ignore_ascii_case("error") {
@@ -3618,7 +3621,7 @@ fn available_space(path: &std::path::Path) -> std::io::Result<u64> {
     fs2::available_space(path)
 }
 
-fn existing_filesystem_probe_path(path: &std::path::Path) -> std::path::PathBuf {
+pub(crate) fn existing_filesystem_probe_path(path: &std::path::Path) -> std::path::PathBuf {
     let mut cursor = if path.as_os_str().is_empty() {
         std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from("."))
     } else {
@@ -4205,2174 +4208,12 @@ fn zccache_json_flag_unsupported(output: &std::process::Output) -> bool {
         || stderr.contains("found argument")
 }
 
-fn clear_zccache_cache() -> Result<(), SoldrError> {
-    let paths = SoldrPaths::new()?;
-    let zccache_dir = managed_zccache_cache_dir(&paths)?;
-    let mut cleared_anything = false;
 
-    if let Some(fetch) = cached_managed_zccache(&paths)? {
-        let _ = run_zccache_command_in_cache_dir(&fetch.binary_path, &["clear"], &zccache_dir)?;
-        println!("cleared zccache artifact cache");
-        cleared_anything = true;
-    }
-    if zccache_dir.exists() {
-        std::fs::remove_dir_all(&zccache_dir)?;
-        println!("removed soldr zccache state dir: {}", zccache_dir.display());
-        cleared_anything = true;
-    }
-    if !cleared_anything {
-        println!(
-            "managed zccache {} not fetched yet",
-            soldr_fetch::MANAGED_ZCCACHE_VERSION
-        );
-    }
-    Ok(())
+pub(crate) struct CommandOutput {
+    pub(crate) stdout: String,
 }
 
-// ---------------------------------------------------------------------------
-// soldr gc — garbage-collect stale Cargo target/ directories.
-// ---------------------------------------------------------------------------
-
-#[derive(Clone, Copy)]
-enum GcMode {
-    Summary,
-    Purge { all: bool },
-}
-
-struct GcInvocation {
-    mode: GcMode,
-    older_than: String,
-    larger_than: String,
-    json: bool,
-}
-
-#[derive(Serialize)]
-struct GcCandidateOutput {
-    path: String,
-    size_bytes: u64,
-    size_human: String,
-    age_seconds: i64,
-    age_human: String,
-    eligible: bool,
-    reason: Option<String>,
-}
-
-#[derive(Serialize)]
-struct GcOutput {
-    schema_version: u32,
-    command: &'static str,
-    mode: &'static str,
-    dry_run: bool,
-    registry_path: String,
-    candidate_count: usize,
-    skipped_count: usize,
-    total_reclaimable_bytes: u64,
-    total_reclaimable_human: String,
-    candidates: Vec<GcCandidateOutput>,
-    largest_candidates: Vec<GcCandidateOutput>,
-    skipped: Vec<GcCandidateOutput>,
-    dropped_missing: usize,
-    deleted_paths: Vec<String>,
-    selected_count: usize,
-    succeeded_count: usize,
-    failed_count: usize,
-    reclaimed_bytes: u64,
-    reclaimed_human: String,
-    error_log_path: Option<String>,
-}
-
-fn run_gc_command(invocation: GcInvocation) -> Result<(), SoldrError> {
-    use soldr_cache::gc::{
-        cleanup_old_gc_logs, parse_duration, parse_size, scan, write_gc_error_log, GcOptions,
-        GcPurgeSummary,
-    };
-
-    let older_than = parse_duration(&invocation.older_than).map_err(SoldrError::Other)?;
-    let larger_than = parse_size(&invocation.larger_than).map_err(SoldrError::Other)?;
-    let purge_all = match invocation.mode {
-        GcMode::Summary => false,
-        GcMode::Purge { all } => all,
-    };
-    let is_summary = matches!(invocation.mode, GcMode::Summary);
-
-    let paths = SoldrPaths::new()?;
-    let dev_roots = resolve_gc_dev_roots(&paths);
-    let db_path = soldr_cache::data_db_path(&paths);
-    let registry = soldr_cache::target_registry::TargetRegistry::open(&db_path)
-        .map_err(|e| SoldrError::Other(format!("failed to open soldr registry: {e}")))?;
-    let gc_log_dir = soldr_cache::gc_log_dir(&paths);
-    cleanup_old_gc_logs(&gc_log_dir)
-        .map_err(|e| SoldrError::Other(format!("failed to clean old gc logs: {e}")))?;
-
-    let options = GcOptions {
-        older_than_seconds: older_than,
-        larger_than_bytes: larger_than,
-        dev_roots,
-        dry_run: is_summary,
-    };
-
-    let report =
-        scan(&registry, &options).map_err(|e| SoldrError::Other(format!("gc scan failed: {e}")))?;
-    let total_reclaimable_bytes = gc_total_reclaimable_bytes(&report.candidates);
-
-    let mut deleted_paths: Vec<String> = Vec::new();
-    let mut purge_summary = GcPurgeSummary::default();
-    let mut error_log_path: Option<std::path::PathBuf> = None;
-
-    if is_summary {
-        if !invocation.json {
-            print_gc_summary(&db_path, &report, total_reclaimable_bytes);
-        }
-    } else if !invocation.json {
-        print_gc_purge_scan(&db_path, &report, total_reclaimable_bytes);
-    }
-
-    if !is_summary {
-        purge_summary =
-            run_gc_purge_candidates(&registry, &report.candidates, purge_all, invocation.json)?;
-        deleted_paths = purge_summary
-            .deleted_paths
-            .iter()
-            .map(|p| p.display().to_string())
-            .collect();
-        if !purge_summary.failures.is_empty() {
-            let args = std::env::args().collect::<Vec<_>>();
-            let path = write_gc_error_log(&gc_log_dir, &args, &purge_summary.failures)
-                .map_err(|e| SoldrError::Other(format!("failed to write gc error log: {e}")))?;
-            error_log_path = Some(path);
-        }
-        if !invocation.json {
-            print_gc_purge_result(&purge_summary, error_log_path.as_deref());
-        }
-    }
-
-    if invocation.json {
-        let output = GcOutput {
-            schema_version: JSON_SCHEMA_VERSION,
-            command: "gc",
-            mode: if is_summary { "summary" } else { "purge" },
-            dry_run: is_summary,
-            registry_path: db_path.display().to_string(),
-            candidate_count: report.candidates.len(),
-            skipped_count: report.skipped.len(),
-            total_reclaimable_bytes,
-            total_reclaimable_human: soldr_cache::target_registry::human_size(
-                total_reclaimable_bytes,
-            ),
-            largest_candidates: gc_largest_candidates(&report.candidates, 5)
-                .into_iter()
-                .map(gc_candidate_output)
-                .collect(),
-            candidates: report
-                .candidates
-                .into_iter()
-                .map(gc_candidate_output)
-                .collect(),
-            skipped: report
-                .skipped
-                .into_iter()
-                .map(gc_candidate_output)
-                .collect(),
-            dropped_missing: report.dropped_missing,
-            deleted_paths,
-            selected_count: purge_summary.selected_count,
-            succeeded_count: purge_summary.succeeded_count,
-            failed_count: purge_summary.failed_count,
-            reclaimed_bytes: purge_summary.reclaimed_bytes,
-            reclaimed_human: soldr_cache::target_registry::human_size(
-                purge_summary.reclaimed_bytes,
-            ),
-            error_log_path: error_log_path.map(|p| p.display().to_string()),
-        };
-        print_json(&output)?;
-    }
-    Ok(())
-}
-
-#[derive(Serialize)]
-struct GcListEntryOutput {
-    path: String,
-    last_used_unix: i64,
-    age_seconds: i64,
-    age_human: String,
-    size_bytes: u64,
-    size_human: String,
-    file_count: u64,
-}
-
-#[derive(Serialize)]
-struct GcListOutput {
-    schema_version: u32,
-    command: &'static str,
-    mode: &'static str,
-    registry_path: String,
-    entry_count: usize,
-    pruned_missing: usize,
-    entries: Vec<GcListEntryOutput>,
-}
-
-fn absolute_path_string(path: &std::path::Path) -> String {
-    std::path::absolute(path)
-        .unwrap_or_else(|_| path.to_path_buf())
-        .display()
-        .to_string()
-}
-
-/// Compute `(size_bytes, file_count)` for a directory using rayon to
-/// fan out across the top-level entries. The per-entry walk is the
-/// existing sequential routine. This keeps the implementation small
-/// while exploiting the typical cargo `target/` layout where the bulk
-/// of bytes sit under a handful of subdirs (`debug/`, `release/`,
-/// per-target triples, etc.).
-fn fast_directory_size_and_files(path: &std::path::Path) -> (u64, u64) {
-    use rayon::prelude::*;
-
-    let metadata = match std::fs::symlink_metadata(path) {
-        Ok(m) => m,
-        Err(_) => return (0, 0),
-    };
-    if metadata.file_type().is_symlink() {
-        return (0, 0);
-    }
-    if metadata.is_file() {
-        return (metadata.len(), 1);
-    }
-    let entries: Vec<std::fs::DirEntry> = match std::fs::read_dir(path) {
-        Ok(iter) => iter.flatten().collect(),
-        Err(_) => return (0, 0),
-    };
-    entries
-        .into_par_iter()
-        .map(|entry| {
-            let entry_path = entry.path();
-            let entry_meta = match entry.metadata() {
-                Ok(m) => m,
-                Err(_) => return (0u64, 0u64),
-            };
-            if entry_meta.file_type().is_symlink() {
-                (0, 0)
-            } else if entry_meta.is_dir() {
-                soldr_cache::target_registry::directory_size_and_files(&entry_path)
-            } else if entry_meta.is_file() {
-                (entry_meta.len(), 1)
-            } else {
-                (0, 0)
-            }
-        })
-        .reduce(
-            || (0u64, 0u64),
-            |a, b| (a.0.saturating_add(b.0), a.1.saturating_add(b.1)),
-        )
-}
-
-fn run_gc_list_command(json: bool) -> Result<(), SoldrError> {
-    use rayon::prelude::*;
-
-    let paths = SoldrPaths::new()?;
-    let db_path = soldr_cache::data_db_path(&paths);
-    let registry = soldr_cache::target_registry::TargetRegistry::open(&db_path)
-        .map_err(|e| SoldrError::Other(format!("failed to open soldr registry: {e}")))?;
-    let rows = registry
-        .list()
-        .map_err(|e| SoldrError::Other(format!("gc list failed: {e}")))?;
-    let now = soldr_cache::target_registry::current_unix_seconds()
-        .map_err(|e| SoldrError::Other(format!("gc list clock error: {e}")))?;
-
-    // Partition rows into those still on disk and those that have
-    // disappeared since the registry was written. Missing rows are
-    // never reported — they're swept out of the registry at the end
-    // via a single batched delete.
-    let (live_rows, missing_paths): (Vec<_>, Vec<_>) = rows.into_par_iter().partition_map(|row| {
-        if row.path.exists() {
-            rayon::iter::Either::Left(row)
-        } else {
-            rayon::iter::Either::Right(row.path)
-        }
-    });
-
-    let entries: Vec<GcListEntryOutput> = live_rows
-        .into_par_iter()
-        .map(|row| {
-            let (size_bytes, file_count) = fast_directory_size_and_files(&row.path);
-            let age_seconds = now.saturating_sub(row.last_used);
-            GcListEntryOutput {
-                path: absolute_path_string(&row.path),
-                last_used_unix: row.last_used,
-                age_seconds,
-                age_human: soldr_cache::target_registry::human_age(age_seconds),
-                size_bytes,
-                size_human: soldr_cache::target_registry::human_size(size_bytes),
-                file_count,
-            }
-        })
-        .collect();
-
-    let pruned_missing = registry
-        .remove_many(&missing_paths)
-        .map_err(|e| SoldrError::Other(format!("failed to prune missing registry rows: {e}")))?;
-
-    if json {
-        let output = GcListOutput {
-            schema_version: JSON_SCHEMA_VERSION,
-            command: "gc",
-            mode: "list",
-            registry_path: db_path.display().to_string(),
-            entry_count: entries.len(),
-            pruned_missing,
-            entries,
-        };
-        print_json(&output)?;
-    } else {
-        println!("soldr gc list: registry: {}", db_path.display());
-        println!(
-            "soldr gc list: {} tracked target dir{}",
-            entries.len(),
-            if entries.len() == 1 { "" } else { "s" }
-        );
-        for entry in &entries {
-            println!(
-                "  {}  size={}  files={}  age={}",
-                entry.path, entry.size_human, entry.file_count, entry.age_human,
-            );
-        }
-        if pruned_missing > 0 {
-            println!(
-                "soldr gc list: pruned {pruned_missing} missing row{} from registry",
-                if pruned_missing == 1 { "" } else { "s" }
-            );
-        }
-    }
-    Ok(())
-}
-
-fn run_gc_purge_candidates(
-    registry: &soldr_cache::target_registry::TargetRegistry,
-    candidates: &[soldr_cache::gc::GcCandidate],
-    purge_all: bool,
-    json: bool,
-) -> Result<soldr_cache::gc::GcPurgeSummary, SoldrError> {
-    let worker_count = gc_purge_worker_count();
-    let (job_tx, job_rx) = mpsc::channel::<soldr_cache::gc::GcCandidate>();
-    let (result_tx, result_rx) = mpsc::channel();
-    let job_rx = Arc::new(Mutex::new(job_rx));
-    let mut workers = Vec::new();
-    for idx in 0..worker_count {
-        let job_rx = Arc::clone(&job_rx);
-        let result_tx = result_tx.clone();
-        let builder = std::thread::Builder::new().name(format!("soldr-gc-{idx}"));
-        workers.push(
-            builder
-                .spawn(move || loop {
-                    let next = {
-                        let rx = job_rx.lock().expect("gc worker channel poisoned");
-                        rx.recv()
-                    };
-                    match next {
-                        Ok(candidate) => {
-                            let _ =
-                                result_tx.send(soldr_cache::gc::delete_candidate_dir(candidate));
-                        }
-                        Err(_) => break,
-                    }
-                })
-                .map_err(|e| SoldrError::Other(format!("failed to start gc worker: {e}")))?,
-        );
-    }
-    drop(result_tx);
-
-    let mut selected_count = 0usize;
-    let mut completed_count = 0usize;
-    let mut outcomes = Vec::new();
-
-    for cand in candidates {
-        let should_delete = purge_all || prompt_gc_purge_candidate(cand);
-        if !should_delete {
-            continue;
-        }
-
-        selected_count += 1;
-        job_tx
-            .send(cand.clone())
-            .map_err(|e| SoldrError::Other(format!("failed to queue gc delete: {e}")))?;
-    }
-    drop(job_tx);
-
-    while completed_count < selected_count {
-        match result_rx.recv_timeout(Duration::from_millis(200)) {
-            Ok(outcome) => {
-                completed_count += 1;
-                outcomes.push(outcome);
-                if !json {
-                    print_gc_purge_progress(completed_count, selected_count);
-                }
-            }
-            Err(mpsc::RecvTimeoutError::Timeout) => {
-                if !json {
-                    print_gc_purge_progress(completed_count, selected_count);
-                }
-            }
-            Err(mpsc::RecvTimeoutError::Disconnected) => break,
-        }
-    }
-
-    for worker in workers {
-        worker
-            .join()
-            .map_err(|_| SoldrError::Other("gc worker panicked".to_string()))?;
-    }
-
-    if !json && selected_count > 0 {
-        eprintln!();
-    }
-
-    soldr_cache::gc::apply_purge_outcomes(registry, outcomes)
-        .map_err(|e| SoldrError::Other(format!("failed to update gc registry: {e}")))
-}
-
-fn gc_purge_worker_count() -> usize {
-    let available = std::thread::available_parallelism()
-        .map(|n| n.get())
-        .unwrap_or(1);
-    gc_purge_worker_count_for(available)
-}
-
-fn gc_purge_worker_count_for(available_parallelism: usize) -> usize {
-    available_parallelism.clamp(1, 4)
-}
-
-fn print_gc_purge_progress(completed: usize, selected: usize) {
-    eprint!("\rsoldr gc purge: deleting selected targets {completed}/{selected}");
-    let _ = std::io::stderr().flush();
-}
-
-fn gc_total_reclaimable_bytes(candidates: &[soldr_cache::gc::GcCandidate]) -> u64 {
-    candidates.iter().map(|c| c.size_bytes).sum()
-}
-
-fn print_gc_summary(
-    db_path: &std::path::Path,
-    report: &soldr_cache::gc::GcReport,
-    total_reclaimable_bytes: u64,
-) {
-    println!("soldr gc: registry: {}", db_path.display());
-    println!(
-        "soldr gc: eligible: {} target dir{}; reclaimable: {}",
-        report.candidates.len(),
-        if report.candidates.len() == 1 {
-            ""
-        } else {
-            "s"
-        },
-        soldr_cache::target_registry::human_size(total_reclaimable_bytes)
-    );
-    println!(
-        "soldr gc: skipped: {}; dropped missing rows: {}",
-        report.skipped.len(),
-        report.dropped_missing
-    );
-
-    if report.candidates.is_empty() {
-        println!("soldr gc: nothing to reclaim.");
-    } else {
-        println!("soldr gc: largest eligible target directories:");
-        for cand in gc_largest_candidates(&report.candidates, 5) {
-            println!(
-                "  {}  size={}  last_used={}",
-                cand.path.display(),
-                soldr_cache::target_registry::human_size(cand.size_bytes),
-                soldr_cache::target_registry::human_age(cand.age_seconds),
-            );
-        }
-        println!("Run 'soldr gc purge' to delete eligible target directories.");
-    }
-}
-
-fn print_gc_purge_scan(
-    db_path: &std::path::Path,
-    report: &soldr_cache::gc::GcReport,
-    total_reclaimable_bytes: u64,
-) {
-    eprintln!(
-        "soldr gc purge: scanned registry at {} ({} candidate dir{}, {} skipped, {} dropped missing, {} reclaimable)",
-        db_path.display(),
-        report.candidates.len(),
-        if report.candidates.len() == 1 { "" } else { "s" },
-        report.skipped.len(),
-        report.dropped_missing,
-        soldr_cache::target_registry::human_size(total_reclaimable_bytes)
-    );
-
-    if report.candidates.is_empty() {
-        eprintln!("soldr gc purge: nothing to delete.");
-    } else {
-        eprintln!("soldr gc purge: candidates");
-        for cand in &report.candidates {
-            eprintln!(
-                "  {}  size={}  age={}",
-                cand.path.display(),
-                soldr_cache::target_registry::human_size(cand.size_bytes),
-                soldr_cache::target_registry::human_age(cand.age_seconds),
-            );
-        }
-    }
-}
-
-fn print_gc_purge_result(
-    summary: &soldr_cache::gc::GcPurgeSummary,
-    error_log_path: Option<&std::path::Path>,
-) {
-    eprintln!(
-        "soldr gc purge: selected {}; succeeded {}; failed {}; reclaimed {}",
-        summary.selected_count,
-        summary.succeeded_count,
-        summary.failed_count,
-        soldr_cache::target_registry::human_size(summary.reclaimed_bytes)
-    );
-    if let Some(path) = error_log_path {
-        eprintln!(
-            "soldr gc purge: detailed deletion errors written to {}",
-            path.display()
-        );
-    }
-}
-
-fn gc_largest_candidates(
-    candidates: &[soldr_cache::gc::GcCandidate],
-    limit: usize,
-) -> Vec<soldr_cache::gc::GcCandidate> {
-    let mut largest = candidates.to_vec();
-    largest.sort_by(|a, b| {
-        b.size_bytes
-            .cmp(&a.size_bytes)
-            .then_with(|| a.path.cmp(&b.path))
-    });
-    largest.truncate(limit);
-    largest
-}
-
-fn gc_candidate_output(c: soldr_cache::gc::GcCandidate) -> GcCandidateOutput {
-    GcCandidateOutput {
-        path: c.path.display().to_string(),
-        size_human: soldr_cache::target_registry::human_size(c.size_bytes),
-        size_bytes: c.size_bytes,
-        age_human: soldr_cache::target_registry::human_age(c.age_seconds),
-        age_seconds: c.age_seconds,
-        eligible: c.eligible,
-        reason: c.reason,
-    }
-}
-
-fn prompt_gc_purge_candidate(cand: &soldr_cache::gc::GcCandidate) -> bool {
-    prompt_yes_no_default_yes(&format!(
-        "soldr gc: delete {} ({}, age {}) ? [Y/n] ",
-        cand.path.display(),
-        soldr_cache::target_registry::human_size(cand.size_bytes),
-        soldr_cache::target_registry::human_age(cand.age_seconds),
-    ))
-}
-
-fn prompt_yes_no_default_yes(prompt: &str) -> bool {
-    use std::io::{BufRead, Write};
-    eprint!("{prompt}");
-    let _ = std::io::stderr().flush();
-    let stdin = std::io::stdin();
-    let mut line = String::new();
-    if stdin.lock().read_line(&mut line).is_err() {
-        return false;
-    }
-    parse_gc_purge_answer(&line)
-}
-
-fn parse_gc_purge_answer(input: &str) -> bool {
-    matches!(input.trim().to_ascii_lowercase().as_str(), "" | "y" | "yes")
-}
-
-/// Resolve the configured `gc.allowlist_roots`, falling back to
-/// `~/dev` when unset.
-fn resolve_gc_dev_roots(paths: &SoldrPaths) -> Vec<std::path::PathBuf> {
-    let config = paths.load_config();
-    let configured = config
-        .gc
-        .allowlist_roots
-        .clone()
-        .unwrap_or_default()
-        .into_iter()
-        .filter(|r| !r.trim().is_empty())
-        .map(|r| soldr_core::expand_user_home(&r))
-        .collect::<Vec<_>>();
-    if !configured.is_empty() {
-        return configured;
-    }
-    if let Ok(home) = soldr_core::user_home_dir() {
-        return vec![home.join("dev")];
-    }
-    Vec::new()
-}
-
-// ---------------------------------------------------------------------------
-// soldr gc cargo / gc locations / gc sweep — issue #323 manual surface.
-// ---------------------------------------------------------------------------
-
-#[derive(Serialize)]
-struct GcCargoOutput {
-    schema_version: u32,
-    command: &'static str,
-    mode: &'static str,
-    toolchain: String,
-    exit_code: i32,
-    dry_run: bool,
-    args: Vec<String>,
-    stdout_bytes: u64,
-    stderr_bytes: u64,
-    skipped: bool,
-    skipped_reason: Option<String>,
-}
-
-#[derive(Serialize)]
-struct GcLocationOutput {
-    kind: &'static str,
-    path: String,
-    exists: bool,
-    size_bytes: u64,
-    size_human: String,
-    file_count: u64,
-    owner: &'static str,
-    purge_safety: &'static str,
-}
-
-#[derive(Serialize)]
-struct GcLocationsOutput {
-    schema_version: u32,
-    command: &'static str,
-    mode: &'static str,
-    locations: Vec<GcLocationOutput>,
-    total_size_bytes: u64,
-    total_size_human: String,
-}
-
-#[derive(Serialize)]
-struct GcSweepOutput {
-    schema_version: u32,
-    command: &'static str,
-    mode: &'static str,
-    dry_run: bool,
-    cargo_gc: Option<GcCargoOutput>,
-    cargo_gc_aggressive: Option<GcCargoOutput>,
-    soldr_targets: Option<SoldrTargetsSummary>,
-    locations: Vec<GcLocationOutput>,
-    elapsed_ms: u128,
-}
-
-#[derive(Serialize, Default)]
-struct SoldrTargetsSummary {
-    selected_count: usize,
-    succeeded_count: usize,
-    failed_count: usize,
-    reclaimed_bytes: u64,
-    reclaimed_human: String,
-}
-
-/// `soldr gc cargo` — shell out to nightly cargo's `-Zgc clean gc`.
-fn run_gc_cargo_command(args: GcCargoArgs) -> Result<(), SoldrError> {
-    let outcome = invoke_cargo_native_gc(&args, false)?;
-    if args.json {
-        print_json(&outcome)?;
-    }
-    if outcome.skipped {
-        // Explicit gc cargo treats a missing nightly as a hard error,
-        // unlike gc sweep which downgrades the missing toolchain to a
-        // skip. We surfaced the skip JSON above for callers that care,
-        // but the exit code must reflect the failure.
-        return Err(SoldrError::Other(
-            outcome
-                .skipped_reason
-                .unwrap_or_else(|| "cargo nightly GC unavailable".into()),
-        ));
-    }
-    if outcome.exit_code != 0 {
-        std::process::exit(outcome.exit_code);
-    }
-    Ok(())
-}
-
-/// Common implementation backing `gc cargo` and the cargo-step of
-/// `gc sweep`. When `skip_when_missing` is true (sweep), a missing
-/// nightly toolchain returns a `skipped = true` outcome with
-/// `exit_code = 0` so the orchestrator can continue.
-fn invoke_cargo_native_gc(
-    args: &GcCargoArgs,
-    skip_when_missing: bool,
-) -> Result<GcCargoOutput, SoldrError> {
-    let toolchain = resolve_gc_cargo_toolchain(args.toolchain.as_deref());
-
-    let mut forwarded: Vec<String> = Vec::new();
-    push_optional_flag(&mut forwarded, "--max-src-age", args.max_src_age.as_deref());
-    push_optional_flag(
-        &mut forwarded,
-        "--max-crate-age",
-        args.max_crate_age.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-index-age",
-        args.max_index_age.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-git-co-age",
-        args.max_git_co_age.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-git-db-age",
-        args.max_git_db_age.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-download-age",
-        args.max_download_age.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-src-size",
-        args.max_src_size.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-crate-size",
-        args.max_crate_size.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-git-size",
-        args.max_git_size.as_deref(),
-    );
-    push_optional_flag(
-        &mut forwarded,
-        "--max-download-size",
-        args.max_download_size.as_deref(),
-    );
-    if args.dry_run {
-        forwarded.push("--dry-run".to_string());
-    }
-
-    // Final `cargo` argv: -Zgc clean gc [forwarded...]
-    let mut cargo_argv: Vec<String> =
-        vec!["-Zgc".to_string(), "clean".to_string(), "gc".to_string()];
-    cargo_argv.extend(forwarded.iter().cloned());
-
-    // We invoke via `rustup run <toolchain> cargo ...` so the
-    // workspace's rust-toolchain.toml override does not silently win.
-    if !rustup_run_available_for(&toolchain) {
-        if skip_when_missing {
-            return Ok(GcCargoOutput {
-                schema_version: JSON_SCHEMA_VERSION,
-                command: "gc",
-                mode: "cargo",
-                toolchain: toolchain.clone(),
-                exit_code: 0,
-                dry_run: args.dry_run,
-                args: cargo_argv,
-                stdout_bytes: 0,
-                stderr_bytes: 0,
-                skipped: true,
-                skipped_reason: Some(format!(
-                    "rustup toolchain {toolchain} not installed; skipping cargo GC"
-                )),
-            });
-        }
-        return Err(SoldrError::Other(format!(
-            "rustup toolchain {toolchain} not installed; install it with `rustup toolchain install {toolchain}` or pass --toolchain <name>"
-        )));
-    }
-
-    let mut command = std::process::Command::new("rustup");
-    command.arg("run").arg(&toolchain).arg("cargo");
-    command.args(&cargo_argv);
-    soldr_core::suppress_windows_console_window(&mut command);
-
-    eprintln!(
-        "soldr gc cargo: rustup run {toolchain} cargo {}",
-        cargo_argv.join(" ")
-    );
-
-    let output = command.output().map_err(|e| {
-        SoldrError::Other(format!(
-            "failed to invoke rustup run {toolchain} cargo: {e}"
-        ))
-    })?;
-
-    // Stream cargo's output through to the user's terminal.
-    use std::io::Write as _;
-    let _ = std::io::stdout().write_all(&output.stdout);
-    let _ = std::io::stderr().write_all(&output.stderr);
-
-    Ok(GcCargoOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "gc",
-        mode: "cargo",
-        toolchain,
-        exit_code: output.status.code().unwrap_or(1),
-        dry_run: args.dry_run,
-        args: cargo_argv,
-        stdout_bytes: output.stdout.len() as u64,
-        stderr_bytes: output.stderr.len() as u64,
-        skipped: false,
-        skipped_reason: None,
-    })
-}
-
-fn push_optional_flag(out: &mut Vec<String>, flag: &str, value: Option<&str>) {
-    if let Some(v) = value {
-        out.push(format!("{flag}={v}"));
-    }
-}
-
-fn resolve_gc_cargo_toolchain(flag: Option<&str>) -> String {
-    flag.map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .or_else(|| {
-            std::env::var(SOLDR_GC_CARGO_TOOLCHAIN_ENV_VAR)
-                .ok()
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty())
-        })
-        .unwrap_or_else(|| "nightly".to_string())
-}
-
-/// Best-effort probe: `rustup toolchain list` must list the supplied
-/// channel. We avoid actually shelling into the toolchain here because
-/// `rustup run <missing> cargo --version` will install on demand,
-/// which we don't want for a probe.
-fn rustup_run_available_for(toolchain: &str) -> bool {
-    let mut command = std::process::Command::new("rustup");
-    command.args(["toolchain", "list"]);
-    soldr_core::suppress_windows_console_window(&mut command);
-    let output = match command.output() {
-        Ok(o) => o,
-        Err(_) => return false,
-    };
-    if !output.status.success() {
-        return false;
-    }
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    stdout
-        .lines()
-        .any(|line| line.trim().starts_with(toolchain))
-}
-
-/// `soldr gc locations` — read-only enumeration of every cache dir
-/// soldr cares about, with sizes (no last-used derivation yet).
-fn run_gc_locations_command(json: bool) -> Result<(), SoldrError> {
-    let paths = SoldrPaths::new()?;
-    let entries = enumerate_cache_locations(&paths);
-    let total_size_bytes: u64 = entries.iter().map(|e| e.size_bytes).sum();
-
-    if json {
-        let output = GcLocationsOutput {
-            schema_version: JSON_SCHEMA_VERSION,
-            command: "gc",
-            mode: "locations",
-            total_size_bytes,
-            total_size_human: soldr_cache::target_registry::human_size(total_size_bytes),
-            locations: entries,
-        };
-        print_json(&output)?;
-    } else {
-        println!("soldr gc locations:");
-        println!(
-            "  total tracked size: {}",
-            soldr_cache::target_registry::human_size(total_size_bytes)
-        );
-        for entry in &entries {
-            println!(
-                "  [{:>8}] {}  size={}  files={}  owner={}  purge_safety={}",
-                entry.kind,
-                entry.path,
-                entry.size_human,
-                entry.file_count,
-                entry.owner,
-                entry.purge_safety,
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Enumerate every directory soldr cares about: cargo home subdirs,
-/// rustup home subdirs, soldr's own cache root, and the state.redb
-/// file. Missing paths are reported with `exists = false` and zero
-/// size so the JSON shape stays predictable.
-fn enumerate_cache_locations(paths: &SoldrPaths) -> Vec<GcLocationOutput> {
-    let mut entries: Vec<GcLocationOutput> = Vec::new();
-
-    if let Some(cargo_home) = soldr_core::resolve_cargo_home() {
-        for (kind, suffix, owner, purge_safety) in &[
-            ("cargo_registry_src", "registry/src", "cargo", "regenerable"),
-            (
-                "cargo_registry_cache",
-                "registry/cache",
-                "cargo",
-                "regenerable",
-            ),
-            (
-                "cargo_registry_index",
-                "registry/index",
-                "cargo",
-                "regenerable",
-            ),
-            ("cargo_git_db", "git/db", "cargo", "regenerable"),
-            (
-                "cargo_git_checkouts",
-                "git/checkouts",
-                "cargo",
-                "regenerable",
-            ),
-        ] {
-            let path = cargo_home.join(suffix);
-            entries.push(gc_location_for(kind, &path, owner, purge_safety));
-        }
-        // .global-cache is a single file in cargo's stable tree.
-        let global_cache = cargo_home.join(".global-cache");
-        entries.push(gc_location_for(
-            "cargo_global_cache",
-            &global_cache,
-            "cargo",
-            "regenerable",
-        ));
-    }
-
-    if let Some(rustup_home) = soldr_core::resolve_rustup_home() {
-        entries.push(gc_location_for(
-            "rustup_toolchains",
-            &rustup_home.join("toolchains"),
-            "rustup",
-            "user_action",
-        ));
-        entries.push(gc_location_for(
-            "rustup_update_hashes",
-            &rustup_home.join("update-hashes"),
-            "rustup",
-            "regenerable",
-        ));
-    }
-
-    entries.push(gc_location_for(
-        "soldr_cache",
-        &paths.cache,
-        "soldr",
-        "regenerable",
-    ));
-    entries.push(gc_location_for(
-        "soldr_state_db",
-        &paths.root.join("state.redb"),
-        "soldr",
-        "user_action",
-    ));
-
-    entries
-}
-
-fn gc_location_for(
-    kind: &'static str,
-    path: &std::path::Path,
-    owner: &'static str,
-    purge_safety: &'static str,
-) -> GcLocationOutput {
-    let exists = path.exists();
-    let (size_bytes, file_count) = if exists {
-        fast_directory_size_and_files(path)
-    } else {
-        (0, 0)
-    };
-    GcLocationOutput {
-        kind,
-        path: path.display().to_string(),
-        exists,
-        size_bytes,
-        size_human: soldr_cache::target_registry::human_size(size_bytes),
-        file_count,
-        owner,
-        purge_safety,
-    }
-}
-
-/// `soldr gc sweep` — orchestrate locations + cargo gc + soldr target
-/// purge in one go.
-fn run_gc_sweep_command(args: GcSweepArgs) -> Result<(), SoldrError> {
-    let start = std::time::Instant::now();
-    let paths = SoldrPaths::new()?;
-    let cargo_gc_enabled = !args.no_cargo_gc;
-
-    // 1. Locations table (always — read-only).
-    let locations = enumerate_cache_locations(&paths);
-
-    // 2. Cargo's clean gc with conservative ages (unless disabled).
-    let cargo_gc_outcome = if cargo_gc_enabled {
-        // Use conservative defaults — let cargo's own ~1mo / ~3mo
-        // policy decide. We don't pass any --max-*-age flags so the
-        // user can configure cargo independently.
-        let cargo_args = GcCargoArgs {
-            dry_run: args.dry_run,
-            toolchain: None,
-            max_src_age: None,
-            max_crate_age: None,
-            max_index_age: None,
-            max_git_co_age: None,
-            max_git_db_age: None,
-            max_download_age: None,
-            max_src_size: None,
-            max_crate_size: None,
-            max_git_size: None,
-            max_download_size: None,
-            json: args.json,
-        };
-        Some(invoke_cargo_native_gc(&cargo_args, true)?)
-    } else {
-        None
-    };
-
-    // 3. soldr's target purge over registered workspaces.
-    let soldr_targets = if args.dry_run {
-        if !args.json {
-            eprintln!("soldr gc sweep: dry-run; skipping soldr target purge");
-        }
-        None
-    } else {
-        Some(run_soldr_target_purge_for_sweep(
-            &paths, args.all, args.json,
-        )?)
-    };
-
-    // 4. Aggressive second cargo pass.
-    let cargo_gc_aggressive = if args.aggressive && cargo_gc_enabled {
-        let cfg = paths.load_config();
-        let floor = cfg.auto_gc.min_age_secs;
-        let aggressive_args = aggressive_cargo_args(args.json, args.dry_run, floor);
-        Some(invoke_cargo_native_gc(&aggressive_args, true)?)
-    } else {
-        None
-    };
-
-    if !args.json {
-        eprintln!("soldr gc sweep: done in {} ms", start.elapsed().as_millis());
-    }
-
-    if args.json {
-        let output = GcSweepOutput {
-            schema_version: JSON_SCHEMA_VERSION,
-            command: "gc",
-            mode: "sweep",
-            dry_run: args.dry_run,
-            cargo_gc: cargo_gc_outcome,
-            cargo_gc_aggressive,
-            soldr_targets,
-            locations,
-            elapsed_ms: start.elapsed().as_millis(),
-        };
-        print_json(&output)?;
-    }
-    Ok(())
-}
-
-fn aggressive_cargo_args(json: bool, dry_run: bool, min_age_secs: u64) -> GcCargoArgs {
-    // Helper: clamp `aggressive_days * 86_400` to the configured min
-    // age. Express the result back in seconds (cargo accepts `s` /
-    // `secs` / `seconds`).
-    let clamp = |days: u64| -> String {
-        let secs =
-            soldr_cache::auto_gc::clamp_age_to_floor(days.saturating_mul(86_400), min_age_secs);
-        format!("{secs}secs")
-    };
-    GcCargoArgs {
-        dry_run,
-        toolchain: None,
-        max_src_age: Some(clamp(7)),
-        max_crate_age: Some(clamp(14)),
-        max_index_age: None,
-        max_git_co_age: Some(clamp(7)),
-        max_git_db_age: None,
-        max_download_age: None,
-        max_src_size: None,
-        max_crate_size: None,
-        max_git_size: None,
-        max_download_size: None,
-        json,
-    }
-}
-
-fn run_soldr_target_purge_for_sweep(
-    _paths: &SoldrPaths,
-    purge_all: bool,
-    json: bool,
-) -> Result<SoldrTargetsSummary, SoldrError> {
-    use soldr_cache::gc::{parse_duration, parse_size, scan, GcOptions};
-    let paths = SoldrPaths::new()?;
-    let dev_roots = resolve_gc_dev_roots(&paths);
-    let db_path = soldr_cache::data_db_path(&paths);
-    let registry = soldr_cache::target_registry::TargetRegistry::open(&db_path)
-        .map_err(|e| SoldrError::Other(format!("failed to open soldr registry: {e}")))?;
-
-    let cfg = paths.load_config();
-    let older_than_seconds = soldr_cache::auto_gc::clamp_age_to_floor(
-        parse_duration("10d").map_err(SoldrError::Other)?,
-        cfg.auto_gc.min_age_secs,
-    );
-    let larger_than_bytes = parse_size("256M").map_err(SoldrError::Other)?;
-
-    let options = GcOptions {
-        older_than_seconds,
-        larger_than_bytes,
-        dev_roots,
-        dry_run: false,
-    };
-    let report =
-        scan(&registry, &options).map_err(|e| SoldrError::Other(format!("gc scan failed: {e}")))?;
-    if report.candidates.is_empty() {
-        return Ok(SoldrTargetsSummary::default());
-    }
-    let purge_summary = run_gc_purge_candidates(&registry, &report.candidates, purge_all, json)?;
-    Ok(SoldrTargetsSummary {
-        selected_count: purge_summary.selected_count,
-        succeeded_count: purge_summary.succeeded_count,
-        failed_count: purge_summary.failed_count,
-        reclaimed_bytes: purge_summary.reclaimed_bytes,
-        reclaimed_human: soldr_cache::target_registry::human_size(purge_summary.reclaimed_bytes),
-    })
-}
-
-// ---------------------------------------------------------------------------
-// Auto-GC under disk pressure (issue #323).
-//
-// Hook lives at the soldr cargo front door. On every cargo invocation
-// the wrapper consults a throttle marker and, if the throttle has
-// expired and the user hasn't opted out, spawns a detached background
-// thread that:
-//
-//   1. enumerates soldr-relevant paths and groups them by volume;
-//   2. probes free space per volume;
-//   3. runs the tiered GC plan only against volumes below the trigger;
-//   4. appends a structured line to ~/.soldr/logs/auto-gc.log.
-//
-// We deliberately spawn instead of running inline so the wrapper never
-// blocks the build. cargo's `.package-cache` mutex handles concurrent
-// invocations of `cargo clean gc` cleanly for us.
-// ---------------------------------------------------------------------------
-
-const AUTO_GC_THROTTLE_SECONDS: u64 = 5 * 60;
-const AUTO_GC_LOG_MAX_BYTES: u64 = 10 * 1024 * 1024;
-const AUTO_GC_DISABLE_ENV_VAR: &str = "SOLDR_AUTO_GC_DISABLED";
-
-fn maybe_kick_auto_gc(paths: &SoldrPaths) {
-    if auto_gc_env_disabled() {
-        return;
-    }
-    let config = paths.load_config().auto_gc;
-    if !config.enabled {
-        return;
-    }
-    let marker = soldr_cache::auto_gc_throttle_marker_path(paths);
-    if !auto_gc_throttle_expired(&marker, AUTO_GC_THROTTLE_SECONDS) {
-        return;
-    }
-    // Touch the marker before spawning so a crashing background thread
-    // doesn't cause us to immediately rerun on the next invocation.
-    let _ = touch_auto_gc_marker(&marker);
-
-    let log_path = soldr_cache::auto_gc_log_path(paths);
-    let paths_root = paths.root.clone();
-    let _ = std::thread::Builder::new()
-        .name("soldr-auto-gc".to_string())
-        .spawn(move || {
-            run_auto_gc_background(paths_root, log_path);
-        });
-}
-
-fn auto_gc_env_disabled() -> bool {
-    match std::env::var(AUTO_GC_DISABLE_ENV_VAR) {
-        Ok(value) => !matches!(
-            value.trim().to_ascii_lowercase().as_str(),
-            "" | "0" | "false" | "no" | "off"
-        ),
-        Err(_) => false,
-    }
-}
-
-fn auto_gc_throttle_expired(marker: &std::path::Path, throttle_seconds: u64) -> bool {
-    let Ok(meta) = std::fs::metadata(marker) else {
-        return true;
-    };
-    let Ok(modified) = meta.modified() else {
-        return true;
-    };
-    let elapsed = std::time::SystemTime::now()
-        .duration_since(modified)
-        .unwrap_or(std::time::Duration::ZERO);
-    elapsed.as_secs() >= throttle_seconds
-}
-
-fn touch_auto_gc_marker(marker: &std::path::Path) -> std::io::Result<()> {
-    if let Some(parent) = marker.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    std::fs::write(marker, "")
-}
-
-fn run_auto_gc_background(paths_root: std::path::PathBuf, log_path: std::path::PathBuf) {
-    use soldr_cache::auto_gc::DiskFreeProbe as _;
-    let start = std::time::Instant::now();
-    let paths = SoldrPaths::with_root(paths_root);
-    let config = paths.load_config().auto_gc;
-    let (validated, warnings) = soldr_cache::auto_gc::validate_config(&config);
-    for warning in &warnings {
-        let _ = append_auto_gc_log_line(&log_path, &format!("warning: {warning}"));
-    }
-
-    let auto_paths = enumerate_auto_gc_paths(&paths);
-    let probe = SystemVolumeProbe;
-    let plans = soldr_cache::auto_gc::plan_auto_gc(&validated, &auto_paths, &probe, &probe);
-    if plans.is_empty() {
-        return; // Either disabled or no volume is below trigger.
-    }
-
-    for plan in &plans {
-        let line = format!(
-            "auto-gc volume={} free_gib={:.2} trigger_gib={} target_gib={} paths={} status=detected",
-            plan.volume_key,
-            (plan.free_bytes as f64) / (soldr_cache::auto_gc::GIB as f64),
-            validated.trigger_free_gb,
-            validated.target_free_gb,
-            plan.paths.len()
-        );
-        let _ = append_auto_gc_log_line(&log_path, &line);
-
-        // Tier 1: conservative cargo GC (no explicit --max-*-age flags
-        // so cargo uses its own conservative defaults). Only attempt
-        // when the volume holds the cargo home.
-        let mut last_tier = 0u8;
-        let cargo_volume_paths = plan
-            .paths
-            .iter()
-            .filter(|p| matches!(p.kind, soldr_cache::auto_gc::AutoGcPathKind::CargoHome))
-            .count();
-        if cargo_volume_paths > 0 {
-            let outcome = run_conservative_cargo_gc_background(&log_path);
-            last_tier = 1;
-            let _ = append_auto_gc_log_line(
-                &log_path,
-                &format!(
-                    "tier=1 volume={} exit_code={} skipped={} reason={}",
-                    plan.volume_key,
-                    outcome.exit_code,
-                    outcome.skipped,
-                    outcome.reason.as_deref().unwrap_or("ran")
-                ),
-            );
-        }
-
-        // Re-probe and decide whether to escalate.
-        let mut free_bytes = probe.free_bytes(&plan.paths[0].path).unwrap_or(0);
-        let target_bytes = validated
-            .target_free_gb
-            .saturating_mul(soldr_cache::auto_gc::GIB);
-
-        // Tier 2: soldr target purge (only if volume holds workspace
-        // targets and we're still under target).
-        if soldr_cache::auto_gc::next_tier(free_bytes, target_bytes, last_tier).is_some() {
-            let workspace_targets: Vec<_> = plan
-                .paths
-                .iter()
-                .filter(|p| {
-                    matches!(
-                        p.kind,
-                        soldr_cache::auto_gc::AutoGcPathKind::WorkspaceTarget
-                    )
-                })
-                .map(|p| p.path.clone())
-                .collect();
-            if !workspace_targets.is_empty() {
-                let reclaimed = run_soldr_target_purge_background(
-                    &paths,
-                    &workspace_targets,
-                    validated.min_age_secs,
-                );
-                last_tier = 2;
-                free_bytes = probe.free_bytes(&plan.paths[0].path).unwrap_or(free_bytes);
-                let _ = append_auto_gc_log_line(
-                    &log_path,
-                    &format!(
-                        "tier=2 volume={} reclaimed_bytes={} free_gib={:.2}",
-                        plan.volume_key,
-                        reclaimed,
-                        (free_bytes as f64) / (soldr_cache::auto_gc::GIB as f64),
-                    ),
-                );
-            }
-        }
-
-        // Tier 3: aggressive cargo GC (clamped to min_age_secs).
-        if soldr_cache::auto_gc::next_tier(free_bytes, target_bytes, last_tier).is_some()
-            && cargo_volume_paths > 0
-        {
-            let ages = soldr_cache::auto_gc::TIER3_AGES.clamped_seconds(validated.min_age_secs);
-            let outcome = run_aggressive_cargo_gc_background(&log_path, &ages);
-            last_tier = 3;
-            free_bytes = probe.free_bytes(&plan.paths[0].path).unwrap_or(free_bytes);
-            let _ = append_auto_gc_log_line(
-                &log_path,
-                &format!(
-                    "tier=3 volume={} exit_code={} skipped={} reason={} free_gib={:.2}",
-                    plan.volume_key,
-                    outcome.exit_code,
-                    outcome.skipped,
-                    outcome.reason.as_deref().unwrap_or("ran"),
-                    (free_bytes as f64) / (soldr_cache::auto_gc::GIB as f64),
-                ),
-            );
-        }
-
-        if soldr_cache::auto_gc::next_tier(free_bytes, target_bytes, last_tier).is_none()
-            && free_bytes < target_bytes
-        {
-            let _ = append_auto_gc_log_line(
-                &log_path,
-                &format!(
-                    "auto-gc warning volume={} free_gib={:.2} target_gib={} \
-                    tiers exhausted; run `soldr gc sweep --aggressive`",
-                    plan.volume_key,
-                    (free_bytes as f64) / (soldr_cache::auto_gc::GIB as f64),
-                    validated.target_free_gb,
-                ),
-            );
-        }
-    }
-
-    let _ = append_auto_gc_log_line(
-        &log_path,
-        &format!(
-            "auto-gc done elapsed_ms={} volumes={}",
-            start.elapsed().as_millis(),
-            plans.len(),
-        ),
-    );
-    let _ = rotate_auto_gc_log_if_needed(&log_path, AUTO_GC_LOG_MAX_BYTES);
-}
-
-struct AutoGcCargoOutcome {
-    exit_code: i32,
-    skipped: bool,
-    reason: Option<String>,
-}
-
-fn run_conservative_cargo_gc_background(log_path: &std::path::Path) -> AutoGcCargoOutcome {
-    let args = GcCargoArgs {
-        dry_run: false,
-        toolchain: None,
-        max_src_age: None,
-        max_crate_age: None,
-        max_index_age: None,
-        max_git_co_age: None,
-        max_git_db_age: None,
-        max_download_age: None,
-        max_src_size: None,
-        max_crate_size: None,
-        max_git_size: None,
-        max_download_size: None,
-        json: true,
-    };
-    match invoke_cargo_native_gc(&args, true) {
-        Ok(outcome) => AutoGcCargoOutcome {
-            exit_code: outcome.exit_code,
-            skipped: outcome.skipped,
-            reason: outcome.skipped_reason,
-        },
-        Err(e) => {
-            let _ = append_auto_gc_log_line(log_path, &format!("tier=1 invoke_error={e}"));
-            AutoGcCargoOutcome {
-                exit_code: 1,
-                skipped: true,
-                reason: Some(format!("invoke_error: {e}")),
-            }
-        }
-    }
-}
-
-fn run_aggressive_cargo_gc_background(
-    log_path: &std::path::Path,
-    ages: &soldr_cache::auto_gc::CargoGcAgeSeconds,
-) -> AutoGcCargoOutcome {
-    let args = GcCargoArgs {
-        dry_run: false,
-        toolchain: None,
-        max_src_age: Some(format!("{}secs", ages.max_src)),
-        max_crate_age: Some(format!("{}secs", ages.max_crate)),
-        max_index_age: None,
-        max_git_co_age: Some(format!("{}secs", ages.max_git_co)),
-        max_git_db_age: None,
-        max_download_age: None,
-        max_src_size: None,
-        max_crate_size: None,
-        max_git_size: None,
-        max_download_size: None,
-        json: true,
-    };
-    match invoke_cargo_native_gc(&args, true) {
-        Ok(outcome) => AutoGcCargoOutcome {
-            exit_code: outcome.exit_code,
-            skipped: outcome.skipped,
-            reason: outcome.skipped_reason,
-        },
-        Err(e) => {
-            let _ = append_auto_gc_log_line(log_path, &format!("tier=3 invoke_error={e}"));
-            AutoGcCargoOutcome {
-                exit_code: 1,
-                skipped: true,
-                reason: Some(format!("invoke_error: {e}")),
-            }
-        }
-    }
-}
-
-fn run_soldr_target_purge_background(
-    paths: &SoldrPaths,
-    workspace_targets: &[std::path::PathBuf],
-    min_age_secs: u64,
-) -> u64 {
-    use soldr_cache::gc::{parse_size, scan, GcOptions};
-    let db_path = soldr_cache::data_db_path(paths);
-    let Ok(registry) = soldr_cache::target_registry::TargetRegistry::open(&db_path) else {
-        return 0;
-    };
-    let larger_than_bytes = parse_size("256M").unwrap_or(256 * 1024 * 1024);
-    // Auto-GC always honors at least the configured min-age floor.
-    // We never go below 1h.
-    let older_than_seconds = soldr_cache::auto_gc::clamp_age_to_floor(min_age_secs, 3600);
-    let options = GcOptions {
-        older_than_seconds,
-        larger_than_bytes,
-        dev_roots: resolve_gc_dev_roots(paths),
-        dry_run: false,
-    };
-    let report = match scan(&registry, &options) {
-        Ok(r) => r,
-        Err(_) => return 0,
-    };
-    // Filter to candidates that actually live on the affected volumes.
-    let mut reclaimed = 0u64;
-    let on_volume: std::collections::HashSet<&std::path::Path> =
-        workspace_targets.iter().map(|p| p.as_path()).collect();
-    for cand in report.candidates {
-        if !on_volume.contains(cand.path.as_path()) {
-            continue;
-        }
-        let bytes = cand.size_bytes;
-        let outcome = soldr_cache::gc::delete_candidate_dir(cand);
-        if outcome.removed {
-            reclaimed = reclaimed.saturating_add(bytes);
-            let _ = registry.remove(&outcome.candidate.path);
-        }
-    }
-    reclaimed
-}
-
-/// Enumerate every soldr-owned path for the auto-GC orchestrator.
-fn enumerate_auto_gc_paths(paths: &SoldrPaths) -> Vec<soldr_cache::auto_gc::AutoGcPath> {
-    let mut out: Vec<soldr_cache::auto_gc::AutoGcPath> = Vec::new();
-    if let Some(cargo_home) = soldr_core::resolve_cargo_home() {
-        out.push(soldr_cache::auto_gc::AutoGcPath {
-            kind: soldr_cache::auto_gc::AutoGcPathKind::CargoHome,
-            path: cargo_home,
-        });
-    }
-    if let Some(rustup_home) = soldr_core::resolve_rustup_home() {
-        out.push(soldr_cache::auto_gc::AutoGcPath {
-            kind: soldr_cache::auto_gc::AutoGcPathKind::RustupHome,
-            path: rustup_home,
-        });
-    }
-    out.push(soldr_cache::auto_gc::AutoGcPath {
-        kind: soldr_cache::auto_gc::AutoGcPathKind::SoldrCache,
-        path: paths.cache.clone(),
-    });
-    let db_path = soldr_cache::data_db_path(paths);
-    if db_path.exists() {
-        if let Ok(registry) = soldr_cache::target_registry::TargetRegistry::open(&db_path) {
-            if let Ok(rows) = registry.list() {
-                for row in rows {
-                    if row.path.exists() {
-                        out.push(soldr_cache::auto_gc::AutoGcPath {
-                            kind: soldr_cache::auto_gc::AutoGcPathKind::WorkspaceTarget,
-                            path: row.path,
-                        });
-                    }
-                }
-            }
-        }
-    }
-    out
-}
-
-/// System volume probe — Windows uses the drive letter (`C`, `D`),
-/// Unix uses the device id from `stat()`. Falls back to the canonical
-/// path's root component when neither is available.
-struct SystemVolumeProbe;
-
-impl soldr_cache::auto_gc::DiskFreeProbe for SystemVolumeProbe {
-    fn free_bytes(&self, path: &std::path::Path) -> Option<u64> {
-        let probe = existing_filesystem_probe_path(path);
-        available_space(&probe).ok()
-    }
-}
-
-impl soldr_cache::auto_gc::VolumeProbe for SystemVolumeProbe {
-    fn volume_key(&self, path: &std::path::Path) -> Option<String> {
-        let probe = existing_filesystem_probe_path(path);
-        volume_key_for_path(&probe)
-    }
-}
-
-#[cfg(windows)]
-fn volume_key_for_path(path: &std::path::Path) -> Option<String> {
-    // On Windows: prefer the canonical path's drive letter (e.g. "C").
-    let canonical = std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
-    let s = canonical.to_string_lossy().to_string();
-    // Strip UNC prefix \\?\ if present.
-    let trimmed = s.trim_start_matches(r"\\?\");
-    let bytes = trimmed.as_bytes();
-    if bytes.len() >= 2 && bytes[1] == b':' && (bytes[0].is_ascii_alphabetic()) {
-        return Some((bytes[0] as char).to_ascii_uppercase().to_string());
-    }
-    None
-}
-
-#[cfg(unix)]
-fn volume_key_for_path(path: &std::path::Path) -> Option<String> {
-    use std::os::unix::fs::MetadataExt;
-    let canonical = std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
-    let meta = std::fs::metadata(&canonical).ok()?;
-    Some(meta.dev().to_string())
-}
-
-fn append_auto_gc_log_line(log_path: &std::path::Path, line: &str) -> std::io::Result<()> {
-    if let Some(parent) = log_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0);
-    let mut file = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(log_path)?;
-    use std::io::Write as _;
-    writeln!(file, "{ts} {line}")?;
-    Ok(())
-}
-
-fn rotate_auto_gc_log_if_needed(log_path: &std::path::Path, max_bytes: u64) -> std::io::Result<()> {
-    let meta = match std::fs::metadata(log_path) {
-        Ok(m) => m,
-        Err(_) => return Ok(()),
-    };
-    if meta.len() < max_bytes {
-        return Ok(());
-    }
-    let archive = log_path.with_extension("log.old");
-    let _ = std::fs::remove_file(&archive);
-    std::fs::rename(log_path, &archive)?;
-    Ok(())
-}
-
-fn emit_startup_target_warning_if_due() {
-    let Ok(paths) = SoldrPaths::new() else { return };
-    let db_path = soldr_cache::data_db_path(&paths);
-    if !db_path.exists() {
-        return;
-    }
-    let Ok(registry) = soldr_cache::target_registry::TargetRegistry::open(&db_path) else {
-        return;
-    };
-    let options = soldr_cache::gc::GcOptions {
-        older_than_seconds: soldr_cache::target_registry::DEFAULT_STALE_AGE_SECONDS,
-        larger_than_bytes: soldr_cache::target_registry::DEFAULT_STALE_SIZE_BYTES,
-        dev_roots: resolve_gc_dev_roots(&paths),
-        dry_run: true,
-    };
-    let marker = soldr_cache::gc_warning_marker_path(&paths);
-    match soldr_cache::gc::maybe_build_startup_warning(&registry, &options, &marker) {
-        Ok(Some(message)) => eprintln!("{message}"),
-        Ok(None) => {}
-        Err(_) => {}
-    }
-}
-
-fn purge_soldr_cache() -> Result<(), SoldrError> {
-    let paths = SoldrPaths::new()?;
-    let mut purged_anything = false;
-
-    purged_anything |= remove_soldr_artifact_dir("cache", &paths.cache)?;
-    purged_anything |= remove_soldr_artifact_dir("bin", &paths.bin)?;
-
-    if !purged_anything {
-        println!("soldr cache is already empty: {}", paths.root.display());
-    }
-
-    Ok(())
-}
-
-fn remove_soldr_artifact_dir(label: &str, path: &std::path::Path) -> Result<bool, SoldrError> {
-    if !path.exists() {
-        return Ok(false);
-    }
-
-    if std::fs::symlink_metadata(path)?.file_type().is_dir() {
-        std::fs::remove_dir_all(path)?;
-        println!("removed soldr {label} dir: {}", path.display());
-    } else {
-        std::fs::remove_file(path)?;
-        println!("removed soldr {label} entry: {}", path.display());
-    }
-    Ok(true)
-}
-
-#[derive(Serialize)]
-struct VersionOutput {
-    schema_version: u32,
-    command: &'static str,
-    soldr_version: String,
-}
-
-#[derive(Serialize)]
-struct StatusOutput {
-    schema_version: u32,
-    command: &'static str,
-    soldr_version: String,
-    target: String,
-    root_dir: String,
-    cache_dir: String,
-    cache_default_enabled: bool,
-    cache_enabled_for_invocation: bool,
-    managed_zccache_version: &'static str,
-    zccache: ZccacheStatusSnapshot,
-}
-
-#[derive(Serialize)]
-struct CacheOutput {
-    schema_version: u32,
-    command: &'static str,
-    soldr_version: String,
-    managed_zccache_version: &'static str,
-    zccache: ZccacheStatusSnapshot,
-}
-
-#[derive(Serialize)]
-struct ZccacheStatusSnapshot {
-    cache_dir: String,
-    state_dir: String,
-    session_log_path: String,
-    session_log_present: bool,
-    journal_path: String,
-    journal_present: bool,
-    session_stats_path: String,
-    session_stats_present: bool,
-    binary_path: Option<String>,
-    binary_fetched: bool,
-    status_lines: Vec<String>,
-    status_empty: bool,
-}
-
-fn version_output() -> VersionOutput {
-    VersionOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "version",
-        soldr_version: soldr_core::version().to_string(),
-    }
-}
-
-fn collect_status_output(cache_enabled: bool) -> Result<StatusOutput, SoldrError> {
-    let target = soldr_core::TargetTriple::detect()?;
-    let paths = SoldrPaths::new()?;
-    Ok(StatusOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "status",
-        soldr_version: soldr_core::version().to_string(),
-        target: target.to_string(),
-        root_dir: paths.root.display().to_string(),
-        cache_dir: paths.cache.display().to_string(),
-        cache_default_enabled: true,
-        cache_enabled_for_invocation: cache_enabled,
-        managed_zccache_version: soldr_fetch::MANAGED_ZCCACHE_VERSION,
-        zccache: collect_zccache_status(&paths)?,
-    })
-}
-
-fn collect_cache_output() -> Result<CacheOutput, SoldrError> {
-    let paths = SoldrPaths::new()?;
-    Ok(CacheOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "cache",
-        soldr_version: soldr_core::version().to_string(),
-        managed_zccache_version: soldr_fetch::MANAGED_ZCCACHE_VERSION,
-        zccache: collect_zccache_status(&paths)?,
-    })
-}
-
-#[derive(Serialize)]
-struct CacheReportOutput {
-    schema_version: u32,
-    command: &'static str,
-    soldr_version: String,
-    managed_zccache_version: &'static str,
-    /// Path to the per-session stats JSON file.
-    session_stats_path: String,
-    /// Whether the session-stats file exists on disk.
-    session_stats_present: bool,
-    /// Path to the per-session JSONL journal.
-    journal_path: String,
-    /// Whether the journal file exists on disk.
-    journal_present: bool,
-    /// Verbatim contents of `last-session-stats.json`, parsed into a JSON
-    /// value. `null` if the file is missing or unparseable.
-    last_session: Option<serde_json::Value>,
-    /// Output of `zccache analyze --json` over the per-session journal,
-    /// when the managed zccache supports it. `null` otherwise.
-    rollups: Option<serde_json::Value>,
-    /// Empty for now — populated by future rule passes that turn the
-    /// session + rollups into AI-readable diagnoses.
-    diagnoses: Vec<serde_json::Value>,
-    /// Why a particular field came back null, when relevant. Each entry
-    /// is a short string the user can search the soldr docs for.
-    notes: Vec<String>,
-}
-
-fn collect_cache_report_output() -> Result<CacheReportOutput, SoldrError> {
-    let paths = SoldrPaths::new()?;
-    let zccache_dir = managed_zccache_cache_dir(&paths)?;
-    let session_stats_path = soldr_cache::session_stats_path(&zccache_dir);
-    let journal_path = soldr_cache::session_journal_path(&zccache_dir);
-    let session_stats_present = session_stats_path.exists();
-    let journal_present = journal_path.exists();
-
-    let mut notes: Vec<String> = Vec::new();
-
-    let last_session = if session_stats_present {
-        match std::fs::read_to_string(&session_stats_path) {
-            Ok(s) => match serde_json::from_str::<serde_json::Value>(s.trim()) {
-                Ok(v) => Some(v),
-                Err(e) => {
-                    notes.push(format!("last_session: unparseable JSON ({e})"));
-                    None
-                }
-            },
-            Err(e) => {
-                notes.push(format!("last_session: read failed ({e})"));
-                None
-            }
-        }
-    } else {
-        notes.push(
-            "last_session: file missing — run a build with managed zccache first".to_string(),
-        );
-        None
-    };
-
-    let rollups = if journal_present {
-        match cached_managed_zccache(&paths)? {
-            Some(fetch) => {
-                let journal_arg = journal_path.display().to_string();
-                let result = run_zccache_command_raw_in_cache_dir(
-                    &fetch.binary_path,
-                    &["analyze", &journal_arg, "--json"],
-                    &zccache_dir,
-                )?;
-                if result.status.success() {
-                    let stdout = String::from_utf8_lossy(&result.stdout);
-                    match serde_json::from_str::<serde_json::Value>(stdout.trim()) {
-                        Ok(v) => Some(v),
-                        Err(e) => {
-                            notes
-                                .push(format!("rollups: zccache analyze stdout unparseable ({e})"));
-                            None
-                        }
-                    }
-                } else if zccache_subcommand_unsupported(&result, "analyze") {
-                    notes.push(format!(
-                        "rollups: managed zccache {} does not yet support `analyze` — upgrade to 1.5.0+",
-                        soldr_fetch::MANAGED_ZCCACHE_VERSION
-                    ));
-                    None
-                } else {
-                    notes.push(format!(
-                        "rollups: zccache analyze exited with status {:?}",
-                        result.status.code()
-                    ));
-                    None
-                }
-            }
-            None => {
-                notes.push(
-                    "rollups: managed zccache binary not yet fetched (no builds run yet)"
-                        .to_string(),
-                );
-                None
-            }
-        }
-    } else {
-        notes
-            .push("rollups: journal missing — soldr writes it on cache-enabled builds".to_string());
-        None
-    };
-
-    Ok(CacheReportOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "cache report",
-        soldr_version: soldr_core::version().to_string(),
-        managed_zccache_version: soldr_fetch::MANAGED_ZCCACHE_VERSION,
-        session_stats_path: session_stats_path.display().to_string(),
-        session_stats_present,
-        journal_path: journal_path.display().to_string(),
-        journal_present,
-        last_session,
-        rollups,
-        diagnoses: Vec::new(),
-        notes,
-    })
-}
-
-/// Heuristic: detect whether a `zccache <subcommand>` invocation failed
-/// because the subcommand does not exist in the running binary (clap
-/// emits "error: unrecognized subcommand"). Used to differentiate
-/// version-skew misses from real failures.
-fn zccache_subcommand_unsupported(output: &std::process::Output, subcommand: &str) -> bool {
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let needles = [
-        "unrecognized subcommand",
-        "unrecognized command",
-        "error: subcommand",
-        "invalid value for",
-    ];
-    let combined = format!("{stderr}\n{stdout}");
-    needles.iter().any(|n| combined.contains(n)) && combined.contains(subcommand)
-}
-
-fn run_cache_prune_target_command(
-    target_dir: std::path::PathBuf,
-    dry_run: bool,
-    json: bool,
-) -> Result<(), SoldrError> {
-    let canonical = std::path::absolute(&target_dir).unwrap_or_else(|_| target_dir.clone());
-    let opts = soldr_cache::prune_target::PruneTargetOptions {
-        target_dir: canonical.clone(),
-        dry_run,
-    };
-    let report = soldr_cache::prune_target::prune_target(&opts)
-        .map_err(|e| SoldrError::Other(format!("cache prune-target failed: {e}")))?;
-
-    if json {
-        let output = build_cache_prune_target_output(&canonical, dry_run, &report);
-        print_json(&output)?;
-    } else {
-        print_cache_prune_target_text(&canonical, dry_run, &report);
-    }
-    Ok(())
-}
-
-fn build_cache_prune_target_output(
-    target_dir: &std::path::Path,
-    dry_run: bool,
-    report: &soldr_cache::prune_target::PruneTargetReport,
-) -> CachePruneTargetOutput {
-    CachePruneTargetOutput {
-        schema_version: JSON_SCHEMA_VERSION,
-        command: "cache prune-target",
-        target_dir: target_dir.display().to_string(),
-        dry_run,
-        scanned: report.scanned,
-        kept: report.kept,
-        deleted: report.deleted,
-        reclaimed_bytes: report.reclaimed_bytes,
-        reclaimed_human: soldr_cache::target_registry::human_size(report.reclaimed_bytes),
-        entries: report
-            .entries
-            .iter()
-            .map(|entry| CachePruneTargetEntryOutput {
-                path: entry.path.display().to_string(),
-                prefix: entry.prefix.clone(),
-                hash: entry.hash.clone(),
-                size_bytes: entry.size_bytes,
-                size_human: soldr_cache::target_registry::human_size(entry.size_bytes),
-                mtime_unix: entry.mtime_unix,
-                action: match entry.action {
-                    soldr_cache::prune_target::PruneAction::Keep => "keep",
-                    soldr_cache::prune_target::PruneAction::Delete => "delete",
-                },
-            })
-            .collect(),
-    }
-}
-
-fn print_cache_prune_target_text(
-    target_dir: &std::path::Path,
-    dry_run: bool,
-    report: &soldr_cache::prune_target::PruneTargetReport,
-) {
-    println!("soldr cache prune-target: {}", target_dir.display());
-    println!(
-        "  mode: {}",
-        if dry_run {
-            "dry-run (use --force to actually delete)"
-        } else {
-            "force"
-        }
-    );
-    println!(
-        "  scanned={} kept={} deleted={} reclaimed={}",
-        report.scanned,
-        report.kept,
-        report.deleted,
-        soldr_cache::target_registry::human_size(report.reclaimed_bytes),
-    );
-    let mut shown = 0usize;
-    for entry in &report.entries {
-        if entry.action != soldr_cache::prune_target::PruneAction::Delete {
-            continue;
-        }
-        if shown == 0 {
-            println!(
-                "  {} entries:",
-                if dry_run { "would delete" } else { "deleted" }
-            );
-        }
-        println!(
-            "    - {} ({})",
-            entry.path.display(),
-            soldr_cache::target_registry::human_size(entry.size_bytes),
-        );
-        shown += 1;
-    }
-    if shown == 0 {
-        println!("  nothing to prune");
-    }
-}
-
-#[derive(Serialize)]
-struct CachePruneTargetOutput {
-    schema_version: u32,
-    command: &'static str,
-    target_dir: String,
-    dry_run: bool,
-    scanned: usize,
-    kept: usize,
-    deleted: usize,
-    reclaimed_bytes: u64,
-    reclaimed_human: String,
-    entries: Vec<CachePruneTargetEntryOutput>,
-}
-
-#[derive(Serialize)]
-struct CachePruneTargetEntryOutput {
-    path: String,
-    prefix: String,
-    hash: String,
-    size_bytes: u64,
-    size_human: String,
-    mtime_unix: i64,
-    action: &'static str,
-}
-
-fn run_cache_report_command(json: bool) -> Result<(), SoldrError> {
-    let output = collect_cache_report_output()?;
-    if json {
-        print_json(&output)?;
-    } else {
-        print_cache_report_output(&output);
-    }
-    Ok(())
-}
-
-fn print_cache_report_output(output: &CacheReportOutput) {
-    println!("soldr cache report");
-    println!(
-        "  session-stats: {} ({})",
-        output.session_stats_path,
-        if output.session_stats_present {
-            "present"
-        } else {
-            "missing"
-        }
-    );
-    println!(
-        "  journal:       {} ({})",
-        output.journal_path,
-        if output.journal_present {
-            "present"
-        } else {
-            "missing"
-        }
-    );
-    if let Some(stats) = &output.last_session {
-        if let Some(rate) = stats.get("hit_rate").and_then(|v| v.as_f64()) {
-            println!("  hit_rate:      {:.1}%", rate * 100.0);
-        }
-        if let Some(hits) = stats.get("hits").and_then(|v| v.as_u64()) {
-            let misses = stats.get("misses").and_then(|v| v.as_u64()).unwrap_or(0);
-            println!("  hits/misses:   {hits}/{misses}");
-        }
-        if let Some(saved_ms) = stats.get("time_saved_ms").and_then(|v| v.as_u64()) {
-            println!("  time_saved:    {saved_ms} ms");
-        }
-    }
-    if let Some(rollups) = &output.rollups {
-        if let Some(by_ext) = rollups.get("by_extension").and_then(|v| v.as_object()) {
-            if !by_ext.is_empty() {
-                println!("  by extension:");
-                for (ext, bucket) in by_ext {
-                    let h = bucket.get("hits").and_then(|v| v.as_u64()).unwrap_or(0);
-                    let m = bucket.get("misses").and_then(|v| v.as_u64()).unwrap_or(0);
-                    println!("    {ext:<14}  hits={h}  misses={m}");
-                }
-            }
-        }
-    }
-    if !output.notes.is_empty() {
-        println!("  notes:");
-        for note in &output.notes {
-            println!("    - {note}");
-        }
-    }
-}
-
-fn collect_zccache_status(paths: &SoldrPaths) -> Result<ZccacheStatusSnapshot, SoldrError> {
-    let zccache_dir = managed_zccache_cache_dir(paths)?;
-    let session_log_path = soldr_cache::session_log_path(&zccache_dir);
-    let session_log_present = session_log_path.exists();
-    let journal_path = soldr_cache::session_journal_path(&zccache_dir);
-    let journal_present = journal_path.exists();
-    let session_stats_path = soldr_cache::session_stats_path(&zccache_dir);
-    let session_stats_present = session_stats_path.exists();
-
-    match cached_managed_zccache(paths)? {
-        Some(fetch) => {
-            let output =
-                run_zccache_command_in_cache_dir(&fetch.binary_path, &["status"], &zccache_dir)?;
-            let stdout = output.stdout.trim();
-            let status_lines = stdout.lines().map(str::to_owned).collect();
-            Ok(ZccacheStatusSnapshot {
-                cache_dir: zccache_dir.display().to_string(),
-                state_dir: zccache_dir.display().to_string(),
-                session_log_path: session_log_path.display().to_string(),
-                session_log_present,
-                journal_path: journal_path.display().to_string(),
-                journal_present,
-                session_stats_path: session_stats_path.display().to_string(),
-                session_stats_present,
-                binary_path: Some(fetch.binary_path.display().to_string()),
-                binary_fetched: true,
-                status_lines,
-                status_empty: stdout.is_empty(),
-            })
-        }
-        None => Ok(ZccacheStatusSnapshot {
-            cache_dir: zccache_dir.display().to_string(),
-            state_dir: zccache_dir.display().to_string(),
-            session_log_path: session_log_path.display().to_string(),
-            session_log_present,
-            journal_path: journal_path.display().to_string(),
-            journal_present,
-            session_stats_path: session_stats_path.display().to_string(),
-            session_stats_present,
-            binary_path: None,
-            binary_fetched: false,
-            status_lines: Vec::new(),
-            status_empty: false,
-        }),
-    }
-}
-
-fn print_status_output(output: &StatusOutput) {
-    println!("soldr {}", output.soldr_version);
-    println!("target: {}", output.target);
-    println!("root dir: {}", output.root_dir);
-    println!("cache dir: {}", output.cache_dir);
-    println!("cache default: enabled");
-    println!(
-        "cache mode: {}",
-        if output.cache_enabled_for_invocation {
-            "enabled"
-        } else {
-            "disabled"
-        }
-    );
-    println!("zccache version: {}", output.managed_zccache_version);
-    print_zccache_status_snapshot(&output.zccache);
-}
-
-fn print_cache_output(output: &CacheOutput) {
-    print_zccache_status_snapshot(&output.zccache);
-}
-
-fn print_zccache_status_snapshot(snapshot: &ZccacheStatusSnapshot) {
-    println!("soldr zccache cache dir: {}", snapshot.cache_dir);
-    println!("soldr zccache state dir: {}", snapshot.state_dir);
-    println!(
-        "last session log: {} ({})",
-        snapshot.session_log_path,
-        if snapshot.session_log_present {
-            "present"
-        } else {
-            "missing"
-        }
-    );
-    println!(
-        "last session journal: {} ({})",
-        snapshot.journal_path,
-        if snapshot.journal_present {
-            "present"
-        } else {
-            "missing"
-        }
-    );
-    println!(
-        "last session stats: {} ({})",
-        snapshot.session_stats_path,
-        if snapshot.session_stats_present {
-            "present"
-        } else {
-            "missing"
-        }
-    );
-
-    if let Some(binary_path) = &snapshot.binary_path {
-        println!("zccache binary: {binary_path}");
-        if snapshot.status_empty {
-            println!("zccache status: no output");
-        } else {
-            for line in &snapshot.status_lines {
-                println!("zccache: {line}");
-            }
-        }
-    } else {
-        println!(
-            "zccache binary: not fetched yet (will fetch managed zccache {} on the first cache-enabled build)",
-            soldr_fetch::MANAGED_ZCCACHE_VERSION
-        );
-    }
-}
-
-fn print_json<T: Serialize>(value: &T) -> Result<(), SoldrError> {
-    serde_json::to_writer_pretty(std::io::stdout(), value)
-        .map_err(|e| SoldrError::Other(format!("failed to serialize JSON output: {e}")))?;
-    println!();
-    Ok(())
-}
-
-struct CommandOutput {
-    stdout: String,
-}
-
-fn managed_zccache_cache_dir(paths: &SoldrPaths) -> Result<std::path::PathBuf, SoldrError> {
+pub(crate) fn managed_zccache_cache_dir(paths: &SoldrPaths) -> Result<std::path::PathBuf, SoldrError> {
     let zccache_dir = normalize_path_for_compare(&soldr_cache::zccache_dir(paths))?;
     let inherited_soldr_managed_dir =
         non_empty_env_path(soldr_cache::MANAGED_ZCCACHE_CACHE_DIR_ENV_VAR)
@@ -6390,7 +4231,7 @@ fn managed_zccache_cache_dir(paths: &SoldrPaths) -> Result<std::path::PathBuf, S
     Ok(zccache_dir)
 }
 
-fn normalize_path_for_compare(path: &std::path::Path) -> Result<std::path::PathBuf, SoldrError> {
+pub(crate) fn normalize_path_for_compare(path: &std::path::Path) -> Result<std::path::PathBuf, SoldrError> {
     if path.is_absolute() {
         Ok(path.to_path_buf())
     } else {
@@ -6398,7 +4239,7 @@ fn normalize_path_for_compare(path: &std::path::Path) -> Result<std::path::PathB
     }
 }
 
-fn run_zccache_command_in_cache_dir(
+pub(crate) fn run_zccache_command_in_cache_dir(
     binary: &std::path::Path,
     args: &[&str],
     cache_dir: &std::path::Path,
@@ -6500,7 +4341,7 @@ fn is_stale_zccache_daemon_start_failure(stderr: &str) -> bool {
         || (stderr.contains("daemon process") && stderr.contains("exists"))
 }
 
-fn run_zccache_command_raw_in_cache_dir(
+pub(crate) fn run_zccache_command_raw_in_cache_dir(
     binary: &std::path::Path,
     args: &[&str],
     cache_dir: &std::path::Path,
@@ -6646,7 +4487,7 @@ async fn fetch_managed_zccache(paths: &SoldrPaths) -> Result<soldr_fetch::FetchR
     soldr_fetch::fetch_zccache_with_paths(paths).await
 }
 
-fn cached_managed_zccache(
+pub(crate) fn cached_managed_zccache(
     paths: &SoldrPaths,
 ) -> Result<Option<soldr_fetch::FetchResult>, SoldrError> {
     if let Some(binary_path) = non_empty_env_path(TEST_ZCCACHE_BIN_ENV_VAR) {
@@ -6661,1715 +4502,4 @@ fn cached_managed_zccache(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        allowed_artifact_classes, build_rust_artifact_plan, build_thin_manifest,
-        cargo_args_specify_target, cargo_args_use_reserved_no_cache,
-        cargo_metadata_passthrough_args, cargo_profile, cargo_target_triple,
-        compute_plan_inputs_hash, dropped_artifact_classes, evaluate_warm_restore_skip,
-        extract_as_pin, first_cargo_subcommand, gc_purge_worker_count_for, is_sccache_wrapper,
-        low_disk_warning_for_free_bytes, low_disk_warning_for_path, normalize_version,
-        parse_gc_purge_answer, parse_rust_artifact_cache_tar_threads, parse_tool_spec,
-        resolve_bundle_walk_thread_count, rustc_wrapper_mode_from_env_var,
-        rustup_resolution_failure, selected_cargo_args, should_self_relocate_for_invocation,
-        should_skip_warm_restore, should_trampoline, stderr_indicates_unknown_session,
-        walk_bundle_files, warm_restore_sentinel_path, warm_restore_skip_enabled,
-        write_thin_manifest, write_warm_restore_sentinel, CargoMetadata, CargoMetadataPackage, Cli,
-        Commands, GcSubcommand, RustArtifactPlan, RustArtifactPlanContext, RustPlanInputs,
-        RustPlanPackages, RustToolchainIdentity, RustcWrapperMode, ThinSliceManifest,
-        WarmRestoreSentinel, WarmRestoreSkipInputs, ZccacheBuildSession, BUNDLE_WALK_THREAD_CAP,
-        LOW_DISK_WARNING_THRESHOLD_BYTES, RUSTC_WRAPPER_OVERRIDE_ENV_VAR,
-        SKIP_WARM_RESTORE_ENV_VAR, THIN_MANIFEST_FILENAME, WARM_RESTORE_MAX_AGE_SECONDS,
-    };
-    use clap::Parser;
-    use soldr_fetch::VersionSpec;
-    use std::ffi::{OsStr, OsString};
-    use std::sync::Mutex;
-    use tempfile::TempDir;
-
-    /// Serialises tests that mutate process-wide environment variables so
-    /// they do not race with each other under parallel `cargo test`. The
-    /// guard objects below restore the previous value on drop, but two
-    /// tests touching the same key concurrently would still observe each
-    /// other's mid-test state without this lock.
-    static ENV_LOCK: Mutex<()> = Mutex::new(());
-
-    /// RAII guard that sets or removes an environment variable for the
-    /// duration of a test and restores the previous value on drop. Modelled
-    /// after the same helper in `soldr-core`'s test module.
-    struct EnvVarGuard {
-        key: &'static str,
-        previous: Option<OsString>,
-    }
-
-    impl EnvVarGuard {
-        fn set(key: &'static str, value: impl AsRef<OsStr>) -> Self {
-            let previous = std::env::var_os(key);
-            std::env::set_var(key, value);
-            Self { key, previous }
-        }
-
-        fn remove(key: &'static str) -> Self {
-            let previous = std::env::var_os(key);
-            std::env::remove_var(key);
-            Self { key, previous }
-        }
-    }
-
-    impl Drop for EnvVarGuard {
-        fn drop(&mut self) {
-            if let Some(value) = &self.previous {
-                std::env::set_var(self.key, value);
-            } else {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
-
-    #[test]
-    fn gc_cli_parses_summary_and_purge_modes() {
-        let summary = Cli::try_parse_from(["soldr", "gc", "--json"]).unwrap();
-        match summary.command {
-            Commands::Gc {
-                command: None,
-                json,
-                ..
-            } => assert!(json, "gc --json should parse as summary JSON"),
-            _ => panic!("expected gc summary command"),
-        }
-
-        let purge = Cli::try_parse_from([
-            "soldr",
-            "gc",
-            "purge",
-            "--all",
-            "--older-than",
-            "30d",
-            "--larger-than",
-            "1GB",
-        ])
-        .unwrap();
-        match purge.command {
-            Commands::Gc {
-                command:
-                    Some(GcSubcommand::Purge {
-                        all,
-                        older_than,
-                        larger_than,
-                        ..
-                    }),
-                ..
-            } => {
-                assert!(all);
-                assert_eq!(older_than, "30d");
-                assert_eq!(larger_than, "1GB");
-            }
-            _ => panic!("expected gc purge command"),
-        }
-    }
-
-    #[test]
-    fn gc_purge_prompt_defaults_enter_to_yes() {
-        for input in ["", "\n", "y", "Y", "yes", " YES "] {
-            assert!(parse_gc_purge_answer(input), "expected {input:?} to accept");
-        }
-        for input in ["n", "no", "anything else"] {
-            assert!(!parse_gc_purge_answer(input), "expected {input:?} to skip");
-        }
-    }
-
-    #[test]
-    fn gc_purge_worker_count_is_bounded() {
-        assert_eq!(gc_purge_worker_count_for(0), 1);
-        assert_eq!(gc_purge_worker_count_for(1), 1);
-        assert_eq!(gc_purge_worker_count_for(2), 2);
-        assert_eq!(gc_purge_worker_count_for(16), 4);
-    }
-
-    #[test]
-    fn low_disk_warning_formats_yellow_below_threshold() {
-        let message = low_disk_warning_for_free_bytes(1536 * 1024 * 1024, true)
-            .expect("expected low-disk warning below threshold");
-        assert!(message.contains("\x1b[33mwarning\x1b[0m"));
-        assert!(message.contains("1.5 GB free"));
-        assert!(message.contains("Run `soldr gc`"));
-    }
-
-    #[test]
-    fn low_disk_warning_omits_at_threshold() {
-        assert!(low_disk_warning_for_free_bytes(LOW_DISK_WARNING_THRESHOLD_BYTES, true).is_none());
-    }
-
-    #[test]
-    fn low_disk_probe_failure_is_nonfatal() {
-        let warning = low_disk_warning_for_path(std::path::Path::new("."), true, |_| {
-            Err(std::io::Error::other("probe failed"))
-        });
-        assert!(warning.is_none());
-    }
-
-    #[test]
-    fn cargo_args_detect_explicit_target_flag() {
-        assert!(cargo_args_specify_target(&[
-            "build".into(),
-            "--target".into(),
-            "x86_64-pc-windows-msvc".into(),
-        ]));
-        assert!(cargo_args_specify_target(&[
-            "build".into(),
-            "--target=x86_64-pc-windows-msvc".into(),
-        ]));
-    }
-
-    #[test]
-    fn cargo_args_ignore_target_after_passthrough_separator() {
-        assert!(!cargo_args_specify_target(&[
-            "test".into(),
-            "--".into(),
-            "--target".into(),
-            "ignored".into(),
-        ]));
-    }
-
-    #[test]
-    fn cargo_args_reject_reserved_no_cache_before_passthrough_separator() {
-        assert!(cargo_args_use_reserved_no_cache(&[
-            "build".into(),
-            "--no-cache".into(),
-        ]));
-        assert!(!cargo_args_use_reserved_no_cache(&[
-            "test".into(),
-            "--".into(),
-            "--no-cache".into(),
-        ]));
-    }
-
-    #[test]
-    fn rustc_wrapper_override_defaults_to_managed_zccache() {
-        assert_eq!(
-            rustc_wrapper_mode_from_env_var(None),
-            RustcWrapperMode::ManagedZccache
-        );
-    }
-
-    #[test]
-    fn rustc_wrapper_override_disables_wrapper_for_empty_or_none() {
-        for value in ["", " ", "none", "NONE"] {
-            assert_eq!(
-                rustc_wrapper_mode_from_env_var(Some(OsStr::new(value))),
-                RustcWrapperMode::Disabled,
-                "expected {value:?} to disable wrapper injection"
-            );
-        }
-    }
-
-    #[test]
-    fn rustc_wrapper_override_uses_custom_wrapper_name() {
-        assert_eq!(
-            rustc_wrapper_mode_from_env_var(Some(OsStr::new("sccache"))),
-            RustcWrapperMode::Custom("sccache".into())
-        );
-    }
-
-    #[test]
-    fn self_relocate_gate_targets_managed_cacheable_cargo_builds() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _wrapper = EnvVarGuard::remove(RUSTC_WRAPPER_OVERRIDE_ENV_VAR);
-
-        assert!(should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "cargo".into(),
-            "build".into(),
-        ]));
-        assert!(should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "--as".into(),
-            env!("CARGO_PKG_VERSION").into(),
-            "cargo".into(),
-            "test".into(),
-        ]));
-        assert!(!should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "cargo".into(),
-            "--version".into(),
-        ]));
-        assert!(!should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "--no-cache".into(),
-            "cargo".into(),
-            "build".into(),
-        ]));
-        assert!(!should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "version".into(),
-        ]));
-
-        let _custom = EnvVarGuard::set(RUSTC_WRAPPER_OVERRIDE_ENV_VAR, "sccache");
-        assert!(!should_self_relocate_for_invocation(&[
-            "soldr".into(),
-            "cargo".into(),
-            "build".into(),
-        ]));
-    }
-
-    #[test]
-    fn sccache_wrapper_detection_accepts_binary_names_and_paths() {
-        assert!(is_sccache_wrapper(OsStr::new("sccache")));
-        assert!(is_sccache_wrapper(OsStr::new("sccache.exe")));
-        assert!(is_sccache_wrapper(OsStr::new("/tmp/tools/sccache")));
-        assert!(!is_sccache_wrapper(OsStr::new("zccache")));
-        assert!(!is_sccache_wrapper(OsStr::new("sccache-proxy")));
-    }
-
-    #[test]
-    fn parse_tool_spec_defaults_to_latest_version() {
-        let (tool, version) = parse_tool_spec("maturin");
-        assert_eq!(tool, "maturin");
-        assert!(matches!(version, VersionSpec::Latest));
-    }
-
-    #[test]
-    fn first_cargo_subcommand_skips_leading_flags() {
-        assert_eq!(
-            first_cargo_subcommand(&["--verbose".into(), "nextest".into(), "run".into()]),
-            Some("nextest")
-        );
-        assert_eq!(
-            first_cargo_subcommand(&["nextest".into(), "run".into()]),
-            Some("nextest")
-        );
-        assert_eq!(first_cargo_subcommand(&["--help".into()]), None);
-        assert_eq!(first_cargo_subcommand(&[]), None);
-    }
-
-    #[test]
-    fn first_cargo_subcommand_stops_at_passthrough_separator() {
-        assert_eq!(
-            first_cargo_subcommand(&["--".into(), "nextest".into()]),
-            None
-        );
-    }
-
-    #[test]
-    fn rust_artifact_plan_selects_external_packages_and_path_exclusions() {
-        let root =
-            std::env::temp_dir().join(format!("soldr-rust-plan-test-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&root);
-        std::fs::create_dir_all(root.join("app/src")).unwrap();
-        std::fs::create_dir_all(root.join("local_dep/src")).unwrap();
-        std::fs::create_dir_all(root.join("target")).unwrap();
-        std::fs::write(root.join("Cargo.lock"), "# lock\n").unwrap();
-        std::fs::write(root.join("Cargo.toml"), "[workspace]\n").unwrap();
-        std::fs::write(root.join("app/Cargo.toml"), "[package]\nname='app'\n").unwrap();
-        std::fs::write(
-            root.join("local_dep/Cargo.toml"),
-            "[package]\nname='local_dep'\n",
-        )
-        .unwrap();
-
-        let metadata = CargoMetadata {
-            workspace_root: root.clone(),
-            target_directory: root.join("target"),
-            workspace_members: vec!["path+file:///repo/app#app@0.1.0".to_string()],
-            packages: vec![
-                CargoMetadataPackage {
-                    id: "path+file:///repo/app#app@0.1.0".to_string(),
-                    source: None,
-                },
-                CargoMetadataPackage {
-                    id: "registry+https://github.com/rust-lang/crates.io-index#serde@1.0.0"
-                        .to_string(),
-                    source: Some("registry+https://github.com/rust-lang/crates.io-index".into()),
-                },
-                CargoMetadataPackage {
-                    id: "path+file:///repo/local_dep#local_dep@0.1.0".to_string(),
-                    source: None,
-                },
-            ],
-        };
-        let toolchain = RustToolchainIdentity {
-            rustc: "rustc 1.0.0-test".to_string(),
-            cargo: "cargo 1.0.0-test".to_string(),
-            channel: "test".to_string(),
-            host: "x86_64-unknown-test".to_string(),
-        };
-        let session = ZccacheBuildSession {
-            binary_path: "zccache".into(),
-            cache_dir: root.join("cache"),
-            session_id: "session-1".to_string(),
-            session_log_path: root.join("cache/logs/last-session.log"),
-            journal_path: root.join("cache/logs/last-session.jsonl"),
-            session_stats_path: root.join("cache/logs/last-session-stats.json"),
-        };
-        let args = vec![
-            "build".to_string(),
-            "--release".to_string(),
-            "--features".to_string(),
-            "serde/derive".to_string(),
-            "--target".to_string(),
-            "x86_64-unknown-linux-gnu".to_string(),
-        ];
-
-        let plan = build_rust_artifact_plan(
-            &metadata,
-            &toolchain,
-            &args,
-            "thin",
-            Some("thin-v1"),
-            &session,
-            None,
-        )
-        .expect("build rust artifact plan");
-
-        assert_eq!(plan.schema_version, 1);
-        assert_eq!(plan.mode, "thin");
-        assert_eq!(plan.cache_profile, Some("thin-v1"));
-        assert_eq!(plan.profile, "release");
-        assert_eq!(plan.target_triple, "x86_64-unknown-linux-gnu");
-        assert_eq!(plan.packages.workspace_package_ids.len(), 1);
-        assert_eq!(plan.packages.selected_package_ids.len(), 1);
-        assert!(plan.packages.selected_package_ids[0].contains("serde"));
-        assert_eq!(plan.packages.excluded_path_package_ids.len(), 1);
-        assert!(plan.allowed_artifact_classes.contains(&"cargo_fingerprint"));
-        assert!(plan.dropped_artifact_classes.is_empty());
-        assert_eq!(plan.cache_schema_version, 1);
-
-        let _ = std::fs::remove_dir_all(&root);
-    }
-
-    #[test]
-    fn rust_artifact_plan_helpers_parse_mode_profile_target_and_metadata_args() {
-        let args = vec![
-            "+stable".to_string(),
-            "build".to_string(),
-            "--locked".to_string(),
-            "--features=fast".to_string(),
-            "--target".to_string(),
-            "wasm32-unknown-unknown".to_string(),
-            "--profile".to_string(),
-            "release-lto".to_string(),
-            "--".to_string(),
-            "--ignored".to_string(),
-        ];
-
-        assert_eq!(cargo_profile(&args), "release-lto");
-        assert_eq!(
-            cargo_target_triple(&args, "x86_64-unknown-linux-gnu"),
-            "wasm32-unknown-unknown"
-        );
-        assert_eq!(
-            selected_cargo_args(&args, &["--features"]),
-            vec!["--features=fast".to_string()]
-        );
-        assert_eq!(allowed_artifact_classes("full", None), Vec::<&str>::new());
-        assert_eq!(
-            cargo_metadata_passthrough_args(&args)
-                .iter()
-                .map(|value| value.to_string_lossy().to_string())
-                .collect::<Vec<_>>(),
-            vec!["--locked".to_string(), "--features=fast".to_string()]
-        );
-    }
-
-    /// `thin-v1` is the legacy slice. It must continue to ship the
-    /// historically-included library-output classes so rollout day 0 is a
-    /// no-op for callers that did not opt in to `thin-v2`.
-    #[test]
-    fn allowed_artifact_classes_thin_v1_keeps_legacy_set() {
-        let allowed = allowed_artifact_classes("thin", Some("thin-v1"));
-        for expected in [
-            "rlib",
-            "rmeta",
-            "dep_info",
-            "proc_macro",
-            "cargo_fingerprint",
-            "build_script_metadata",
-            "build_script_output",
-        ] {
-            assert!(
-                allowed.contains(&expected),
-                "thin-v1 must keep {expected} in the allowlist; got {allowed:?}"
-            );
-        }
-        assert!(dropped_artifact_classes("thin", Some("thin-v1")).is_empty());
-    }
-
-    /// `thin-v2` aggressively prunes the slice. The categories listed in
-    /// `docs/THIN_TARGET_CACHE_PRUNING.md` Section 3.2 must NOT appear in the
-    /// allowlist, and the new fingerprint split (`cargo_fingerprint_meta`,
-    /// dropping `cargo_fingerprint_outputs`) must be honored.
-    #[test]
-    fn allowed_artifact_classes_thin_v2_drops_heavy_categories() {
-        let allowed = allowed_artifact_classes("thin", Some("thin-v2"));
-
-        // Drop list per design Section 3.2.
-        for forbidden in [
-            "rlib",
-            "rmeta",
-            "proc_macro",
-            "incremental",
-            "cargo_fingerprint",
-            "cargo_fingerprint_outputs",
-            "build_script_build",
-            "dwo",
-            "pdb",
-            "dsym",
-        ] {
-            assert!(
-                !allowed.contains(&forbidden),
-                "thin-v2 must drop {forbidden} from the allowlist; got {allowed:?}"
-            );
-        }
-
-        // Keep list per design Section 3.1.
-        for required in [
-            "cargo_fingerprint_meta",
-            "dep_info",
-            "build_script_metadata",
-            "build_script_output",
-        ] {
-            assert!(
-                allowed.contains(&required),
-                "thin-v2 must keep {required} in the allowlist; got {allowed:?}"
-            );
-        }
-
-        // The drop list is surfaced as data so zccache can short-circuit.
-        let dropped = dropped_artifact_classes("thin", Some("thin-v2"));
-        for forbidden in [
-            "incremental",
-            "rlib",
-            "rmeta",
-            "proc_macro",
-            "build_script_build",
-            "dwo",
-            "pdb",
-            "dsym",
-            "cargo_fingerprint_outputs",
-        ] {
-            assert!(
-                dropped.contains(&forbidden),
-                "thin-v2 must publish {forbidden} in dropped_artifact_classes; got {dropped:?}"
-            );
-        }
-    }
-
-    /// Bumping `cache_schema_version` from 1 to 2 is the contract zccache
-    /// uses to decide whether the new fingerprint split is in effect.
-    #[test]
-    fn rust_artifact_plan_bumps_cache_schema_version_for_thin_v2() {
-        let root = std::env::temp_dir().join(format!(
-            "soldr-rust-plan-thinv2-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap()
-                .subsec_nanos(),
-        ));
-        let _ = std::fs::remove_dir_all(&root);
-        std::fs::create_dir_all(root.join("app/src")).unwrap();
-        std::fs::create_dir_all(root.join("target")).unwrap();
-        std::fs::write(root.join("Cargo.lock"), "# lock\n").unwrap();
-        std::fs::write(root.join("Cargo.toml"), "[workspace]\n").unwrap();
-        std::fs::write(root.join("app/Cargo.toml"), "[package]\nname='app'\n").unwrap();
-
-        let metadata = CargoMetadata {
-            workspace_root: root.clone(),
-            target_directory: root.join("target"),
-            workspace_members: vec!["path+file:///repo/app#app@0.1.0".to_string()],
-            packages: vec![CargoMetadataPackage {
-                id: "path+file:///repo/app#app@0.1.0".to_string(),
-                source: None,
-            }],
-        };
-        let toolchain = RustToolchainIdentity {
-            rustc: "rustc 1.0.0-test".to_string(),
-            cargo: "cargo 1.0.0-test".to_string(),
-            channel: "test".to_string(),
-            host: "x86_64-unknown-test".to_string(),
-        };
-        let session = ZccacheBuildSession {
-            binary_path: "zccache".into(),
-            cache_dir: root.join("cache"),
-            session_id: "session-thinv2".to_string(),
-            session_log_path: root.join("cache/logs/last-session.log"),
-            journal_path: root.join("cache/logs/last-session.jsonl"),
-            session_stats_path: root.join("cache/logs/last-session-stats.json"),
-        };
-
-        let plan = build_rust_artifact_plan(
-            &metadata,
-            &toolchain,
-            &["build".to_string()],
-            "thin",
-            Some("thin-v2"),
-            &session,
-            None,
-        )
-        .expect("build rust artifact plan");
-
-        assert_eq!(plan.schema_version, 1, "outer schema is unchanged");
-        assert_eq!(
-            plan.cache_schema_version, 2,
-            "thin-v2 bumps the cache-side schema so zccache can branch on it"
-        );
-        assert_eq!(plan.cache_profile, Some("thin-v2"));
-        assert!(plan.allowed_artifact_classes.contains(&"dep_info"));
-        assert!(!plan.allowed_artifact_classes.contains(&"rlib"));
-
-        let _ = std::fs::remove_dir_all(&root);
-    }
-
-    /// The manifest must enumerate every regular file in the bundle, with
-    /// relative POSIX-style paths and either a size or `null`. It must NOT
-    /// list directories or its own filename.
-    #[test]
-    fn thin_manifest_enumerates_only_files_actually_present() {
-        let bundle = tempfile::tempdir().expect("tempdir for bundle");
-        let bundle_path = bundle.path();
-
-        // Build a representative bundle layout: nested dir + a file at root +
-        // an empty subdir (must not appear in the manifest).
-        std::fs::create_dir_all(bundle_path.join("debug/.fingerprint/foo-abc")).unwrap();
-        std::fs::create_dir_all(bundle_path.join("debug/deps")).unwrap();
-        std::fs::create_dir_all(bundle_path.join("debug/empty_subdir")).unwrap();
-        std::fs::write(
-            bundle_path.join("debug/.fingerprint/foo-abc/invoked.timestamp"),
-            "",
-        )
-        .unwrap();
-        std::fs::write(
-            bundle_path.join("debug/.fingerprint/foo-abc/dep-lib-foo"),
-            b"abc123",
-        )
-        .unwrap();
-        std::fs::write(bundle_path.join("debug/deps/foo-abc.d"), b"foo.rs:\n").unwrap();
-        std::fs::write(bundle_path.join("CACHEDIR.TAG"), b"Signature: 8a4773\n").unwrap();
-
-        let manifest = build_thin_manifest(bundle_path, "thin-v2").expect("build manifest");
-
-        assert_eq!(manifest.schema_version, 2);
-        assert_eq!(manifest.cache_profile, "thin-v2");
-
-        let paths: Vec<&str> = manifest.files.iter().map(|f| f.path.as_str()).collect();
-        // Sorted, POSIX-style, no manifest self-reference, no empty dir.
-        assert_eq!(
-            paths,
-            vec![
-                "CACHEDIR.TAG",
-                "debug/.fingerprint/foo-abc/dep-lib-foo",
-                "debug/.fingerprint/foo-abc/invoked.timestamp",
-                "debug/deps/foo-abc.d",
-            ],
-        );
-        // Sizes are populated for files that exist on disk.
-        let by_path: std::collections::HashMap<_, _> = manifest
-            .files
-            .iter()
-            .map(|f| (f.path.as_str(), f.size_bytes))
-            .collect();
-        assert_eq!(
-            by_path.get("debug/.fingerprint/foo-abc/dep-lib-foo"),
-            Some(&Some(6))
-        );
-        assert_eq!(
-            by_path.get("debug/.fingerprint/foo-abc/invoked.timestamp"),
-            Some(&Some(0))
-        );
-    }
-
-    /// The on-disk manifest emitted by `write_thin_manifest` must round-trip
-    /// through serde so downstream verifiers can deserialize it without
-    /// surprises (no field renames, no missing fields).
-    #[test]
-    fn thin_manifest_round_trips_through_serde() {
-        let bundle = tempfile::tempdir().expect("tempdir for manifest round-trip");
-        let bundle_path = bundle.path();
-        std::fs::create_dir_all(bundle_path.join("debug/deps")).unwrap();
-        std::fs::write(
-            bundle_path.join("debug/deps/example.d"),
-            b"example: src/lib.rs\n",
-        )
-        .unwrap();
-
-        write_thin_manifest(bundle_path, Some("thin-v2")).expect("write manifest");
-
-        let manifest_path = bundle_path.join(THIN_MANIFEST_FILENAME);
-        assert!(
-            manifest_path.is_file(),
-            "manifest must land at the well-known path"
-        );
-
-        let raw = std::fs::read_to_string(&manifest_path).expect("read manifest");
-        let parsed: ThinSliceManifest = serde_json::from_str(&raw).expect("deserialize manifest");
-
-        assert_eq!(parsed.schema_version, 2);
-        assert_eq!(parsed.cache_profile, "thin-v2");
-        assert_eq!(parsed.files.len(), 1);
-        assert_eq!(parsed.files[0].path, "debug/deps/example.d");
-
-        // Serializing the parsed value back must produce a JSON document that
-        // deserializes to an equal value (canonical round-trip).
-        let serialized = serde_json::to_string(&parsed).expect("serialize manifest");
-        let reparsed: ThinSliceManifest =
-            serde_json::from_str(&serialized).expect("re-deserialize manifest");
-        assert_eq!(parsed, reparsed);
-    }
-
-    /// A second `write_thin_manifest` call into the same bundle directory
-    /// must not list the previously-written manifest among its own entries.
-    #[test]
-    fn thin_manifest_does_not_self_reference_on_repeat_save() {
-        let bundle = tempfile::tempdir().expect("tempdir for repeat save");
-        let bundle_path = bundle.path();
-        std::fs::write(bundle_path.join("only.txt"), b"hello").unwrap();
-
-        write_thin_manifest(bundle_path, Some("thin-v2")).expect("first manifest write");
-        write_thin_manifest(bundle_path, Some("thin-v2")).expect("second manifest write");
-
-        let raw = std::fs::read_to_string(bundle_path.join(THIN_MANIFEST_FILENAME))
-            .expect("read manifest");
-        let parsed: ThinSliceManifest = serde_json::from_str(&raw).expect("parse manifest");
-
-        let paths: Vec<&str> = parsed.files.iter().map(|f| f.path.as_str()).collect();
-        assert_eq!(paths, vec!["only.txt"]);
-    }
-
-    #[test]
-    fn known_subcommand_registry_recognizes_phase_two_tools() {
-        for sub in ["nextest", "deny", "audit", "llvm-cov"] {
-            let spec = soldr_fetch::lookup_by_cargo_subcommand(sub)
-                .unwrap_or_else(|| panic!("missing registry entry for cargo {sub}"));
-            assert_eq!(spec.cargo_subcommand, Some(sub));
-            assert!(spec.crate_name.starts_with("cargo-"));
-        }
-    }
-
-    #[test]
-    fn known_subcommand_registry_recognizes_phase_three_tools() {
-        for sub in ["udeps", "semver-checks", "expand", "watch"] {
-            let spec = soldr_fetch::lookup_by_cargo_subcommand(sub)
-                .unwrap_or_else(|| panic!("missing registry entry for cargo {sub}"));
-            assert_eq!(spec.cargo_subcommand, Some(sub));
-            assert!(spec.crate_name.starts_with("cargo-"));
-        }
-    }
-
-    #[test]
-    fn top_level_tools_are_not_cargo_subcommands() {
-        for crate_name in [
-            "cross",
-            "mdbook",
-            "cbindgen",
-            "wasm-pack",
-            "trunk",
-            "sccache",
-        ] {
-            let spec = soldr_fetch::lookup_by_crate(crate_name)
-                .unwrap_or_else(|| panic!("missing registry entry for {crate_name}"));
-            assert_eq!(spec.cargo_subcommand, None);
-        }
-    }
-
-    #[test]
-    fn soldr_itself_is_registered_for_self_trampoline() {
-        let spec = soldr_fetch::lookup_by_crate("soldr")
-            .expect("soldr should be registered in known_tools for --as trampoline");
-        assert_eq!(spec.binary_name, "soldr");
-        assert_eq!(spec.repo, Some(("zackees", "soldr")));
-        assert_eq!(spec.cargo_subcommand, None);
-    }
-
-    #[test]
-    fn extract_as_pin_extracts_space_separated_flag_before_subcommand() {
-        let (version, rest) = extract_as_pin(&[
-            "--as".into(),
-            "0.5.2".into(),
-            "cargo".into(),
-            "build".into(),
-        ])
-        .unwrap();
-        assert_eq!(version, Some("0.5.2".into()));
-        assert_eq!(rest, vec!["cargo".to_string(), "build".into()]);
-    }
-
-    #[test]
-    fn extract_as_pin_extracts_equals_form() {
-        let (version, rest) =
-            extract_as_pin(&["--as=0.5.2".into(), "cargo".into(), "build".into()]).unwrap();
-        assert_eq!(version, Some("0.5.2".into()));
-        assert_eq!(rest, vec!["cargo".to_string(), "build".into()]);
-    }
-
-    #[test]
-    fn extract_as_pin_preserves_other_leading_flags() {
-        let (version, rest) = extract_as_pin(&[
-            "--no-cache".into(),
-            "--as".into(),
-            "0.5.2".into(),
-            "cargo".into(),
-        ])
-        .unwrap();
-        assert_eq!(version, Some("0.5.2".into()));
-        assert_eq!(rest, vec!["--no-cache".to_string(), "cargo".into()]);
-    }
-
-    #[test]
-    fn extract_as_pin_ignores_flag_after_subcommand() {
-        let args = vec!["cargo".into(), "--as".into(), "0.5.2".into()];
-        let (version, rest) = extract_as_pin(&args).unwrap();
-        assert_eq!(version, None);
-        assert_eq!(rest, args);
-    }
-
-    #[test]
-    fn extract_as_pin_ignores_flag_after_passthrough_separator() {
-        let args = vec!["cargo".into(), "--".into(), "--as".into(), "0.5.2".into()];
-        let (version, rest) = extract_as_pin(&args).unwrap();
-        assert_eq!(version, None);
-        assert_eq!(rest, args);
-    }
-
-    #[test]
-    fn extract_as_pin_rejects_missing_value() {
-        let err = extract_as_pin(&["--as".into()]).unwrap_err();
-        assert!(err.to_string().contains("requires a version"));
-    }
-
-    #[test]
-    fn extract_as_pin_rejects_empty_value() {
-        let err = extract_as_pin(&["--as".into(), "".into()]).unwrap_err();
-        assert!(err.to_string().contains("must not be empty"));
-        let err2 = extract_as_pin(&["--as=".into()]).unwrap_err();
-        assert!(err2.to_string().contains("requires a version"));
-    }
-
-    #[test]
-    fn extract_as_pin_rejects_duplicate_flag() {
-        let err =
-            extract_as_pin(&["--as".into(), "0.5.2".into(), "--as=0.4.0".into()]).unwrap_err();
-        assert!(err.to_string().contains("more than once"));
-    }
-
-    #[test]
-    fn normalize_version_strips_leading_v() {
-        assert_eq!(normalize_version("0.5.2"), "0.5.2");
-        assert_eq!(normalize_version("v0.5.2"), "0.5.2");
-        assert_eq!(normalize_version("  v0.5.2 "), "0.5.2");
-    }
-
-    #[test]
-    fn should_trampoline_matches_current_version_as_no_op() {
-        assert!(!should_trampoline(env!("CARGO_PKG_VERSION")));
-        assert!(!should_trampoline(&format!(
-            "v{}",
-            env!("CARGO_PKG_VERSION")
-        )));
-        assert!(should_trampoline("0.0.0-not-this-version"));
-    }
-
-    #[test]
-    fn rustup_resolution_failure_appends_ci_guidance() {
-        let error = rustup_resolution_failure(
-            "rustc",
-            b"error: toolchain '1.94.1-x86_64-pc-windows-msvc' is not installed",
-        );
-
-        let rendered = error.to_string();
-        assert!(rendered.contains("failed to resolve rustc via rustup: error: toolchain '1.94.1-x86_64-pc-windows-msvc' is not installed"));
-        assert!(rendered.contains("pins Rust in rust-toolchain.toml"));
-        assert!(rendered.contains("generic stable toolchain"));
-        assert!(rendered.contains("RUSTUP_TOOLCHAIN"));
-        assert!(rendered.contains("setup-soldr action path"));
-    }
-
-    /// Regression test for the zccache v1.4.0 wire-compat bug. zccache
-    /// v1.4.0 deserializes the plan with `#[serde(deny_unknown_fields)]`
-    /// and does NOT know about `cache_profile` / `dropped_artifact_classes`.
-    /// Therefore the default `thin-v1` (and `full`) JSON must look exactly
-    /// like the pre-PR plan: neither field may appear in the JSON. The
-    /// thin-v2 opt-in is allowed (and required) to surface them.
-    #[test]
-    fn rust_artifact_plan_thin_v1_json_omits_new_fields_for_zccache_compat() {
-        let plan = RustArtifactPlan {
-            schema_version: 1,
-            mode: "thin".to_string(),
-            cache_profile: Some("thin-v1"),
-            workspace_root: "/tmp/ws".to_string(),
-            target_dir: "/tmp/ws/target".to_string(),
-            toolchain: RustToolchainIdentity {
-                rustc: "rustc 1.0.0".to_string(),
-                cargo: "cargo 1.0.0".to_string(),
-                channel: "stable".to_string(),
-                host: "x86_64-unknown-linux-gnu".to_string(),
-            },
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            profile: "release".to_string(),
-            inputs: RustPlanInputs {
-                features_hash: "f".to_string(),
-                rustflags_hash: "r".to_string(),
-                env_hash: "e".to_string(),
-                lockfile_hash: "l".to_string(),
-                cargo_config_hash: "c".to_string(),
-                manifest_hashes: vec![],
-            },
-            packages: RustPlanPackages {
-                selected_package_ids: vec![],
-                workspace_package_ids: vec![],
-                excluded_path_package_ids: vec![],
-            },
-            allowed_artifact_classes: vec!["cargo_fingerprint"],
-            dropped_artifact_classes: vec![],
-            cache_schema_version: 1,
-            journal_log_path: None,
-        };
-
-        let json = serde_json::to_string(&plan).expect("serialize thin-v1 plan");
-        assert!(
-            !json.contains("\"cache_profile\""),
-            "thin-v1 plan must NOT serialize cache_profile (zccache v1.4.0 \
-             rejects unknown fields); got: {json}"
-        );
-        assert!(
-            !json.contains("\"dropped_artifact_classes\""),
-            "thin-v1 plan must NOT serialize dropped_artifact_classes; got: {json}"
-        );
-    }
-
-    /// `full` mode also predates the new fields and zccache's strict
-    /// deserializer rejects them, so `cache_profile == None` plus an empty
-    /// drop list must serialize without either field.
-    #[test]
-    fn rust_artifact_plan_full_mode_json_omits_new_fields() {
-        let plan = RustArtifactPlan {
-            schema_version: 1,
-            mode: "full".to_string(),
-            cache_profile: None,
-            workspace_root: "/tmp/ws".to_string(),
-            target_dir: "/tmp/ws/target".to_string(),
-            toolchain: RustToolchainIdentity {
-                rustc: "rustc 1.0.0".to_string(),
-                cargo: "cargo 1.0.0".to_string(),
-                channel: "stable".to_string(),
-                host: "x86_64-unknown-linux-gnu".to_string(),
-            },
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            profile: "release".to_string(),
-            inputs: RustPlanInputs {
-                features_hash: "f".to_string(),
-                rustflags_hash: "r".to_string(),
-                env_hash: "e".to_string(),
-                lockfile_hash: "l".to_string(),
-                cargo_config_hash: "c".to_string(),
-                manifest_hashes: vec![],
-            },
-            packages: RustPlanPackages {
-                selected_package_ids: vec![],
-                workspace_package_ids: vec![],
-                excluded_path_package_ids: vec![],
-            },
-            allowed_artifact_classes: vec![],
-            dropped_artifact_classes: vec![],
-            cache_schema_version: 1,
-            journal_log_path: None,
-        };
-
-        let json = serde_json::to_string(&plan).expect("serialize full plan");
-        assert!(!json.contains("\"cache_profile\""), "got: {json}");
-        assert!(
-            !json.contains("\"dropped_artifact_classes\""),
-            "got: {json}"
-        );
-    }
-
-    /// thin-v2 is the opt-in that ships the new wire fields. zccache
-    /// builds that consume thin-v2 must see both `cache_profile` and the
-    /// non-empty `dropped_artifact_classes` list.
-    #[test]
-    fn rust_artifact_plan_thin_v2_json_includes_new_fields() {
-        let plan = RustArtifactPlan {
-            schema_version: 1,
-            mode: "thin".to_string(),
-            cache_profile: Some("thin-v2"),
-            workspace_root: "/tmp/ws".to_string(),
-            target_dir: "/tmp/ws/target".to_string(),
-            toolchain: RustToolchainIdentity {
-                rustc: "rustc 1.0.0".to_string(),
-                cargo: "cargo 1.0.0".to_string(),
-                channel: "stable".to_string(),
-                host: "x86_64-unknown-linux-gnu".to_string(),
-            },
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            profile: "release".to_string(),
-            inputs: RustPlanInputs {
-                features_hash: "f".to_string(),
-                rustflags_hash: "r".to_string(),
-                env_hash: "e".to_string(),
-                lockfile_hash: "l".to_string(),
-                cargo_config_hash: "c".to_string(),
-                manifest_hashes: vec![],
-            },
-            packages: RustPlanPackages {
-                selected_package_ids: vec![],
-                workspace_package_ids: vec![],
-                excluded_path_package_ids: vec![],
-            },
-            allowed_artifact_classes: vec!["dep_info"],
-            dropped_artifact_classes: vec!["rlib", "rmeta"],
-            cache_schema_version: 2,
-            journal_log_path: None,
-        };
-
-        let json = serde_json::to_string(&plan).expect("serialize thin-v2 plan");
-        assert!(
-            json.contains("\"cache_profile\":\"thin-v2\""),
-            "thin-v2 must serialize cache_profile; got: {json}"
-        );
-        assert!(
-            json.contains("\"dropped_artifact_classes\""),
-            "thin-v2 must serialize dropped_artifact_classes; got: {json}"
-        );
-    }
-
-    fn warm_restore_test_plan() -> RustArtifactPlan {
-        RustArtifactPlan {
-            schema_version: 1,
-            mode: "thin".to_string(),
-            cache_profile: Some("thin-v1"),
-            workspace_root: "/tmp/ws".to_string(),
-            target_dir: "/tmp/ws/target".to_string(),
-            toolchain: RustToolchainIdentity {
-                rustc: "rustc 1.0.0-test".to_string(),
-                cargo: "cargo 1.0.0-test".to_string(),
-                channel: "stable".to_string(),
-                host: "x86_64-unknown-linux-gnu".to_string(),
-            },
-            target_triple: "x86_64-unknown-linux-gnu".to_string(),
-            profile: "test".to_string(),
-            inputs: RustPlanInputs {
-                features_hash: "F".to_string(),
-                rustflags_hash: "R".to_string(),
-                env_hash: "E".to_string(),
-                lockfile_hash: "L".to_string(),
-                cargo_config_hash: "C".to_string(),
-                manifest_hashes: vec!["M1".to_string(), "M2".to_string()],
-            },
-            packages: RustPlanPackages {
-                selected_package_ids: vec!["serde@1.0.0".to_string()],
-                workspace_package_ids: vec!["app@0.1.0".to_string()],
-                excluded_path_package_ids: vec![],
-            },
-            allowed_artifact_classes: vec!["rlib", "rmeta"],
-            dropped_artifact_classes: vec![],
-            cache_schema_version: 1,
-            journal_log_path: Some("/tmp/journal".to_string()),
-        }
-    }
-
-    fn warm_restore_test_sentinel(plan: &RustArtifactPlan) -> WarmRestoreSentinel {
-        WarmRestoreSentinel {
-            schema_version: 1,
-            plan_inputs_hash: compute_plan_inputs_hash(plan),
-            target_dir: plan.target_dir.clone(),
-            github_run_id: "111".to_string(),
-            github_job: "test".to_string(),
-            github_run_attempt: "1".to_string(),
-            session_id: "session-1".to_string(),
-            saved_at_unix_seconds: 1_000_000,
-        }
-    }
-
-    /// The sentinel hash must change whenever any plan input cargo would
-    /// consult to decide freshness changes. Otherwise the warm-restore
-    /// short-circuit could fire across step pairs that are not actually
-    /// equivalent.
-    #[test]
-    fn plan_inputs_hash_changes_when_inputs_change() {
-        let plan_a = warm_restore_test_plan();
-        let mut plan_b = warm_restore_test_plan();
-        plan_b.inputs.lockfile_hash = "different".to_string();
-        assert_ne!(
-            compute_plan_inputs_hash(&plan_a),
-            compute_plan_inputs_hash(&plan_b),
-        );
-
-        let mut plan_c = warm_restore_test_plan();
-        plan_c.toolchain.rustc = "rustc 9.9.9".to_string();
-        assert_ne!(
-            compute_plan_inputs_hash(&plan_a),
-            compute_plan_inputs_hash(&plan_c),
-        );
-
-        let mut plan_d = warm_restore_test_plan();
-        plan_d.target_triple = "aarch64-apple-darwin".to_string();
-        assert_ne!(
-            compute_plan_inputs_hash(&plan_a),
-            compute_plan_inputs_hash(&plan_d),
-        );
-    }
-
-    /// Cosmetic plan fields (the journal path, the schema version we
-    /// already pin to 1) must not leak into the sentinel hash, so an
-    /// unrelated path swap does not invalidate the warm-restore optim.
-    #[test]
-    fn plan_inputs_hash_ignores_cosmetic_fields() {
-        let plan_a = warm_restore_test_plan();
-        let mut plan_b = warm_restore_test_plan();
-        plan_b.journal_log_path = Some("/tmp/other-journal".to_string());
-        plan_b.workspace_root = "/different/ws".to_string();
-        assert_eq!(
-            compute_plan_inputs_hash(&plan_a),
-            compute_plan_inputs_hash(&plan_b),
-        );
-    }
-
-    /// Happy path: sentinel proves the same plan was just saved into the
-    /// same target dir from the same CI job/attempt — restore is skipped.
-    #[test]
-    fn warm_restore_skip_fires_on_exact_match() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_some(), "expected skip; got {result:?}");
-    }
-
-    /// Plain "no sentinel on disk" must fall through to the normal restore.
-    #[test]
-    fn warm_restore_skip_falls_through_when_sentinel_missing() {
-        let plan = warm_restore_test_plan();
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: "111",
-            github_job: "test",
-            github_run_attempt: "1",
-            now_unix_seconds: 1_000_000,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        assert!(evaluate_warm_restore_skip(None, &inputs).is_none());
-    }
-
-    /// Sentinel from a prior re-run attempt must NOT short-circuit into a
-    /// fresh attempt — the action restored the cache from scratch and the
-    /// `target/` mtimes are no longer guaranteed to be the live ones.
-    #[test]
-    fn warm_restore_skip_rejects_mismatched_run_attempt() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: "2", // different attempt
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// Sentinel from a different job in the same workflow must not bleed
-    /// across job boundaries even when the run id matches.
-    #[test]
-    fn warm_restore_skip_rejects_mismatched_job() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: "other-job",
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// Sentinel for an unrelated target dir (e.g. a sibling workspace
-    /// also writing into the shared bundle dir) must not short-circuit.
-    #[test]
-    fn warm_restore_skip_rejects_mismatched_target_dir() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: "/tmp/different-target",
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// Once a plan input changes (lockfile bump, new manifest, etc.) the
-    /// sentinel hash diverges and restore must run normally.
-    #[test]
-    fn warm_restore_skip_rejects_mismatched_inputs_hash() {
-        let plan = warm_restore_test_plan();
-        let mut sentinel = warm_restore_test_sentinel(&plan);
-        sentinel.plan_inputs_hash = "stale-hash".to_string();
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// Stale sentinels (older than the configured window) must not
-    /// short-circuit. Otherwise a leftover sentinel from a previous
-    /// workflow run could cause skipping in a fresh job that happened to
-    /// inherit the same env identifiers.
-    #[test]
-    fn warm_restore_skip_rejects_stale_sentinel() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let now = sentinel.saved_at_unix_seconds + WARM_RESTORE_MAX_AGE_SECONDS + 1;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// A future-version sentinel (say after a soldr upgrade that bumps
-    /// the schema) must be ignored, never crash, and force a normal
-    /// restore on the next invocation.
-    #[test]
-    fn warm_restore_skip_rejects_unknown_schema_version() {
-        let plan = warm_restore_test_plan();
-        let mut sentinel = warm_restore_test_sentinel(&plan);
-        sentinel.schema_version = 99;
-        let now = sentinel.saved_at_unix_seconds + 60;
-        let inputs_hash = compute_plan_inputs_hash(&plan);
-        let inputs = WarmRestoreSkipInputs {
-            plan_inputs_hash: &inputs_hash,
-            plan_target_dir: &plan.target_dir,
-            github_run_id: &sentinel.github_run_id,
-            github_job: &sentinel.github_job,
-            github_run_attempt: &sentinel.github_run_attempt,
-            now_unix_seconds: now,
-            max_age_seconds: WARM_RESTORE_MAX_AGE_SECONDS,
-        };
-        let result = evaluate_warm_restore_skip(Some(&sentinel), &inputs);
-        assert!(result.is_none());
-    }
-
-    /// Sentinel must round-trip as JSON without dropping fields, so
-    /// disk-roundtrip behavior is observable here too (the
-    /// filesystem-bound caller relies on serde to be exact).
-    #[test]
-    fn warm_restore_sentinel_round_trips_json() {
-        let plan = warm_restore_test_plan();
-        let sentinel = warm_restore_test_sentinel(&plan);
-        let json = serde_json::to_string(&sentinel).expect("serialize sentinel");
-        let parsed: WarmRestoreSentinel = serde_json::from_str(&json).expect("parse sentinel back");
-        assert_eq!(parsed, sentinel);
-    }
-
-    /// Build a `RustArtifactPlanContext` whose plan-derived fields match
-    /// `plan` and whose filesystem-touching paths live under `tempdir`. The
-    /// other fields are filled with deterministic placeholders so tests can
-    /// inspect them without caring about the daemon plumbing they would
-    /// drive in production.
-    fn warm_restore_test_context(
-        plan: &RustArtifactPlan,
-        tempdir: &TempDir,
-    ) -> RustArtifactPlanContext {
-        let root = tempdir.path();
-        RustArtifactPlanContext {
-            path: root.join("plan.json"),
-            zccache_binary: root.join("zccache"),
-            cache_dir: root.join("cache"),
-            zccache_daemon_cache_dir: root.join("daemon"),
-            session_id: "session-test".to_string(),
-            journal_path: root.join("journal"),
-            backend: "fs".to_string(),
-            cache_profile: Some("thin-v1"),
-            plan_inputs_hash: compute_plan_inputs_hash(plan),
-            target_dir: plan.target_dir.clone(),
-        }
-    }
-
-    /// With the gating env var enabled, `write_warm_restore_sentinel` must
-    /// materialise a JSON sentinel under the plan's cache dir whose fields
-    /// reflect the plan inputs and the current GitHub Actions env. This is
-    /// the producer half of the warm-restore short-circuit.
-    #[test]
-    fn write_warm_restore_sentinel_emits_matching_json_when_enabled() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "1");
-        let _run = EnvVarGuard::set("GITHUB_RUN_ID", "run-42");
-        let _job = EnvVarGuard::set("GITHUB_JOB", "test-job");
-        let _attempt = EnvVarGuard::set("GITHUB_RUN_ATTEMPT", "3");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-
-        write_warm_restore_sentinel(&ctx);
-
-        let sentinel_path = warm_restore_sentinel_path(&ctx);
-        let raw = std::fs::read_to_string(&sentinel_path)
-            .expect("sentinel file should exist after write");
-        let sentinel: WarmRestoreSentinel =
-            serde_json::from_str(&raw).expect("sentinel JSON should parse");
-
-        assert_eq!(sentinel.schema_version, 1);
-        assert_eq!(sentinel.plan_inputs_hash, ctx.plan_inputs_hash);
-        assert_eq!(sentinel.target_dir, ctx.target_dir);
-        assert_eq!(sentinel.github_run_id, "run-42");
-        assert_eq!(sentinel.github_job, "test-job");
-        assert_eq!(sentinel.github_run_attempt, "3");
-        assert_eq!(sentinel.session_id, ctx.session_id);
-    }
-
-    /// When the gating env var is explicitly opted out (falsy value), the
-    /// producer must be a strict no-op so the short-circuit cannot
-    /// accidentally fire on the next invocation. No sentinel file should
-    /// appear on disk.
-    #[test]
-    fn write_warm_restore_sentinel_is_noop_when_disabled() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "0");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-
-        write_warm_restore_sentinel(&ctx);
-
-        let sentinel_path = warm_restore_sentinel_path(&ctx);
-        assert!(
-            !sentinel_path.exists(),
-            "no sentinel should be written when {SKIP_WARM_RESTORE_ENV_VAR} is set to a falsy value"
-        );
-    }
-
-    /// Full filesystem round-trip: write a sentinel that exactly matches
-    /// the current plan and CI env, then ask `should_skip_warm_restore`
-    /// whether it should fire. The short-circuit must return `Some` with
-    /// a non-empty operator-visible reason string.
-    #[test]
-    fn should_skip_warm_restore_returns_some_on_full_match() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "1");
-        let _run = EnvVarGuard::set("GITHUB_RUN_ID", "run-7");
-        let _job = EnvVarGuard::set("GITHUB_JOB", "build");
-        let _attempt = EnvVarGuard::set("GITHUB_RUN_ATTEMPT", "1");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-        let sentinel_path = warm_restore_sentinel_path(&ctx);
-        std::fs::create_dir_all(sentinel_path.parent().expect("sentinel has parent dir"))
-            .expect("create sentinel parent");
-        let sentinel = WarmRestoreSentinel {
-            schema_version: 1,
-            plan_inputs_hash: ctx.plan_inputs_hash.clone(),
-            target_dir: ctx.target_dir.clone(),
-            github_run_id: "run-7".to_string(),
-            github_job: "build".to_string(),
-            github_run_attempt: "1".to_string(),
-            session_id: "session-prev".to_string(),
-            saved_at_unix_seconds: super::current_unix_seconds(),
-        };
-        std::fs::write(
-            &sentinel_path,
-            serde_json::to_string(&sentinel).expect("serialize sentinel"),
-        )
-        .expect("write sentinel");
-
-        let result = should_skip_warm_restore(&ctx);
-        let reason = result.expect("expected Some(reason) on full match");
-        assert!(
-            !reason.is_empty(),
-            "skip reason should be non-empty for operator visibility"
-        );
-    }
-
-    /// A sentinel left behind by a previous invocation with a different
-    /// `plan_inputs_hash` (e.g. after a lockfile bump) must not fire the
-    /// short-circuit even when the file is otherwise present and fresh.
-    #[test]
-    fn should_skip_warm_restore_returns_none_on_hash_mismatch() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "1");
-        let _run = EnvVarGuard::set("GITHUB_RUN_ID", "run-7");
-        let _job = EnvVarGuard::set("GITHUB_JOB", "build");
-        let _attempt = EnvVarGuard::set("GITHUB_RUN_ATTEMPT", "1");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-        let sentinel_path = warm_restore_sentinel_path(&ctx);
-        std::fs::create_dir_all(sentinel_path.parent().expect("sentinel has parent dir"))
-            .expect("create sentinel parent");
-        let sentinel = WarmRestoreSentinel {
-            schema_version: 1,
-            plan_inputs_hash: "stale-hash-from-previous-step".to_string(),
-            target_dir: ctx.target_dir.clone(),
-            github_run_id: "run-7".to_string(),
-            github_job: "build".to_string(),
-            github_run_attempt: "1".to_string(),
-            session_id: "session-prev".to_string(),
-            saved_at_unix_seconds: super::current_unix_seconds(),
-        };
-        std::fs::write(
-            &sentinel_path,
-            serde_json::to_string(&sentinel).expect("serialize sentinel"),
-        )
-        .expect("write sentinel");
-
-        assert!(should_skip_warm_restore(&ctx).is_none());
-    }
-
-    /// When no sentinel file exists at all, the short-circuit must fall
-    /// through without panicking on the missing-file IO error.
-    #[test]
-    fn should_skip_warm_restore_returns_none_when_sentinel_missing() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "1");
-        let _run = EnvVarGuard::set("GITHUB_RUN_ID", "run-7");
-        let _job = EnvVarGuard::set("GITHUB_JOB", "build");
-        let _attempt = EnvVarGuard::set("GITHUB_RUN_ATTEMPT", "1");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-        assert!(!warm_restore_sentinel_path(&ctx).exists());
-
-        assert!(should_skip_warm_restore(&ctx).is_none());
-    }
-
-    /// With the gating env var explicitly opted out (`"0"`), the
-    /// short-circuit must stay off even when a perfectly-matching sentinel
-    /// exists. This is the safety property that lets operators disable the
-    /// feature on demand without having to clear stale sentinel files.
-    #[test]
-    fn should_skip_warm_restore_returns_none_when_disabled_even_with_match() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, "0");
-        let _run = EnvVarGuard::set("GITHUB_RUN_ID", "run-7");
-        let _job = EnvVarGuard::set("GITHUB_JOB", "build");
-        let _attempt = EnvVarGuard::set("GITHUB_RUN_ATTEMPT", "1");
-
-        let tempdir = TempDir::new().expect("create tempdir");
-        let plan = warm_restore_test_plan();
-        let ctx = warm_restore_test_context(&plan, &tempdir);
-        let sentinel_path = warm_restore_sentinel_path(&ctx);
-        std::fs::create_dir_all(sentinel_path.parent().expect("sentinel has parent dir"))
-            .expect("create sentinel parent");
-        let sentinel = WarmRestoreSentinel {
-            schema_version: 1,
-            plan_inputs_hash: ctx.plan_inputs_hash.clone(),
-            target_dir: ctx.target_dir.clone(),
-            github_run_id: "run-7".to_string(),
-            github_job: "build".to_string(),
-            github_run_attempt: "1".to_string(),
-            session_id: "session-prev".to_string(),
-            saved_at_unix_seconds: super::current_unix_seconds(),
-        };
-        std::fs::write(
-            &sentinel_path,
-            serde_json::to_string(&sentinel).expect("serialize sentinel"),
-        )
-        .expect("write sentinel");
-
-        assert!(should_skip_warm_restore(&ctx).is_none());
-    }
-
-    /// After the #229 validation flip, an unset env var must enable the
-    /// short-circuit by default. This locks in the default-on contract so
-    /// future refactors cannot regress it without updating the test.
-    #[test]
-    fn warm_restore_skip_enabled_defaults_on() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-        let _skip = EnvVarGuard::remove(SKIP_WARM_RESTORE_ENV_VAR);
-
-        assert!(
-            warm_restore_skip_enabled(),
-            "warm-restore skip must default to enabled when {SKIP_WARM_RESTORE_ENV_VAR} is unset"
-        );
-    }
-
-    /// The default-on flip preserves an explicit opt-out path: each of the
-    /// recognised falsy spellings (`0`, `false`, `no`, `off`, empty string,
-    /// case-insensitive) must disable the short-circuit.
-    #[test]
-    fn warm_restore_skip_enabled_respects_explicit_falsy() {
-        let _lock = ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-
-        for value in ["0", "false", "FALSE", "No", "off", "OFF", "", "  0  "] {
-            let _skip = EnvVarGuard::set(SKIP_WARM_RESTORE_ENV_VAR, value);
-            assert!(
-                !warm_restore_skip_enabled(),
-                "warm-restore skip must be disabled when {SKIP_WARM_RESTORE_ENV_VAR} is set to {value:?}"
-            );
-        }
-    }
-
-    // -------- stderr_indicates_unknown_session (issue #265) --------
-
-    #[test]
-    fn unknown_session_detector_rejects_empty_stderr() {
-        assert!(!stderr_indicates_unknown_session(b""));
-    }
-
-    #[test]
-    fn unknown_session_detector_matches_exact_zccache_line() {
-        let stderr = b"zccache error: unknown session: abc-123\n";
-        assert!(stderr_indicates_unknown_session(stderr));
-    }
-
-    #[test]
-    fn unknown_session_detector_matches_substring_mid_line() {
-        // The marker can appear anywhere in the stream, not necessarily at
-        // the start of a line.
-        let stderr = b"prelude blah blah unknown session: 0000 trailing\n";
-        assert!(stderr_indicates_unknown_session(stderr));
-    }
-
-    #[test]
-    fn unknown_session_detector_ignores_unrelated_session_mentions() {
-        // The word "session" alone is not enough; we only treat the literal
-        // "unknown session:" marker as a resync trigger.
-        let stderr = b"zccache info: session started\nzccache info: session ok\n";
-        assert!(!stderr_indicates_unknown_session(stderr));
-    }
-
-    #[test]
-    fn unknown_session_detector_tolerates_non_utf8_bytes() {
-        // Surround the marker with raw non-UTF-8 byte sequences; the
-        // detector must not panic and must still find the literal needle.
-        let mut stderr: Vec<u8> = vec![0xFF, 0xFE, 0x80, 0x81];
-        stderr.extend_from_slice(b"zccache error: unknown session: deadbeef\n");
-        stderr.extend_from_slice(&[0xC3, 0x28, 0xA0]);
-        assert!(stderr_indicates_unknown_session(&stderr));
-    }
-
-    #[test]
-    fn unknown_session_detector_rejects_partial_marker() {
-        // "unknown sessio" (missing the trailing "n:") must NOT match — we
-        // only resync on the exact daemon-emitted marker.
-        let stderr = b"unknown sessio\n";
-        assert!(!stderr_indicates_unknown_session(stderr));
-    }
-
-    #[test]
-    fn tar_threads_unset_or_blank_yields_none() {
-        assert!(parse_rust_artifact_cache_tar_threads("").unwrap().is_none());
-        assert!(parse_rust_artifact_cache_tar_threads("   ")
-            .unwrap()
-            .is_none());
-    }
-
-    #[test]
-    fn tar_threads_auto_is_normalized_lowercase() {
-        assert_eq!(
-            parse_rust_artifact_cache_tar_threads("auto").unwrap(),
-            Some("auto".to_string())
-        );
-        assert_eq!(
-            parse_rust_artifact_cache_tar_threads("  AUTO ").unwrap(),
-            Some("auto".to_string())
-        );
-    }
-
-    #[test]
-    fn tar_threads_positive_integer_passes_through() {
-        for raw in ["1", "4", "8", "16"] {
-            assert_eq!(
-                parse_rust_artifact_cache_tar_threads(raw).unwrap(),
-                Some(raw.to_string())
-            );
-        }
-    }
-
-    #[test]
-    fn tar_threads_rejects_zero_negative_and_garbage() {
-        for raw in ["0", "-1", "1.5", "twelve", "auto4", "4 threads"] {
-            let err = parse_rust_artifact_cache_tar_threads(raw)
-                .expect_err(&format!("expected error for {raw:?}"));
-            let msg = err.to_string();
-            assert!(
-                msg.contains("SOLDR_TARGET_CACHE_TAR_THREADS"),
-                "error for {raw:?} must mention the env var, got {msg}"
-            );
-        }
-    }
-
-    /// Unset / `auto` / case-variants of `auto` must all yield `None`, which
-    /// signals "use rayon's global thread pool" to `walk_bundle_files`.
-    #[test]
-    fn bundle_walk_thread_count_auto_yields_none() {
-        for raw in ["", "  ", "auto", "AUTO", " Auto "] {
-            assert_eq!(
-                resolve_bundle_walk_thread_count(raw).unwrap(),
-                None,
-                "raw {raw:?} should resolve to None (auto)"
-            );
-        }
-    }
-
-    /// An explicit `1` must turn into `Some(1)` so the walk takes the
-    /// sequential fallback path (no rayon overhead).
-    #[test]
-    fn bundle_walk_thread_count_one_forces_sequential() {
-        assert_eq!(resolve_bundle_walk_thread_count("1").unwrap(), Some(1));
-    }
-
-    /// In-range explicit counts pass through unmodified; values above the
-    /// internal cap are clamped down to `BUNDLE_WALK_THREAD_CAP`.
-    #[test]
-    fn bundle_walk_thread_count_clamps_to_cap() {
-        assert_eq!(resolve_bundle_walk_thread_count("2").unwrap(), Some(2));
-        assert_eq!(
-            resolve_bundle_walk_thread_count("8").unwrap(),
-            Some(BUNDLE_WALK_THREAD_CAP)
-        );
-        // 64 → capped at BUNDLE_WALK_THREAD_CAP.
-        assert_eq!(
-            resolve_bundle_walk_thread_count("64").unwrap(),
-            Some(BUNDLE_WALK_THREAD_CAP)
-        );
-        assert_eq!(
-            resolve_bundle_walk_thread_count("9999").unwrap(),
-            Some(BUNDLE_WALK_THREAD_CAP)
-        );
-    }
-
-    /// Garbage values inherited from the parser must still propagate as
-    /// errors here so callers on the bare `RUSTC_WRAPPER` passthrough path
-    /// (which bypasses the cargo front-door validation) get a clear message
-    /// instead of a silent default.
-    #[test]
-    fn bundle_walk_thread_count_rejects_garbage() {
-        for raw in ["0", "twelve", "1.5"] {
-            let err = resolve_bundle_walk_thread_count(raw)
-                .expect_err(&format!("expected error for {raw:?}"));
-            assert!(
-                err.to_string().contains("SOLDR_TARGET_CACHE_TAR_THREADS"),
-                "error must reference the env var name"
-            );
-        }
-    }
-
-    /// Build a bundle layout with a handful of files at varying depths and
-    /// verify that the walker returns one entry per regular file with the
-    /// correct relative path string (forward-slashed, root-relative).
-    fn populate_walk_bundle_fixture(root: &std::path::Path) {
-        std::fs::create_dir_all(root.join("debug/deps")).unwrap();
-        std::fs::create_dir_all(root.join("debug/build")).unwrap();
-        std::fs::write(root.join("debug/deps/a.rlib"), b"alpha").unwrap();
-        std::fs::write(root.join("debug/deps/b.rmeta"), b"beta!!").unwrap();
-        std::fs::write(root.join("debug/build/c.txt"), b"gamma").unwrap();
-        std::fs::write(root.join("top.txt"), b"delta-delta").unwrap();
-    }
-
-    /// The sequential path (`Some(1)`) must enumerate every file with the
-    /// expected relative paths and sizes. This is the baseline against which
-    /// the parallel walks are compared for determinism.
-    #[test]
-    fn walk_bundle_files_sequential_lists_every_file_with_size() {
-        let bundle = tempfile::tempdir().expect("tempdir");
-        populate_walk_bundle_fixture(bundle.path());
-
-        let mut entries =
-            walk_bundle_files(bundle.path(), Some(1)).expect("sequential walk must succeed");
-        entries.sort_by(|a, b| a.path.cmp(&b.path));
-
-        let observed: Vec<_> = entries
-            .iter()
-            .map(|e| (e.path.as_str(), e.size_bytes))
-            .collect();
-        assert_eq!(
-            observed,
-            vec![
-                ("debug/build/c.txt", Some(5)),
-                ("debug/deps/a.rlib", Some(5)),
-                ("debug/deps/b.rmeta", Some(6)),
-                ("top.txt", Some(11)),
-            ]
-        );
-    }
-
-    /// Output of the walk must be byte-identical (after the caller's
-    /// canonical sort) regardless of whether the metadata phase ran
-    /// sequentially, on rayon's global pool, or on a scoped explicit pool.
-    /// This is the determinism acceptance criterion from issue #272.
-    #[test]
-    fn walk_bundle_files_parallel_matches_sequential_after_sort() {
-        let bundle = tempfile::tempdir().expect("tempdir");
-        populate_walk_bundle_fixture(bundle.path());
-
-        let mut sequential =
-            walk_bundle_files(bundle.path(), Some(1)).expect("sequential walk must succeed");
-        sequential.sort_by(|a, b| a.path.cmp(&b.path));
-
-        for thread_count in [None, Some(2), Some(BUNDLE_WALK_THREAD_CAP)] {
-            let mut parallel = walk_bundle_files(bundle.path(), thread_count)
-                .unwrap_or_else(|e| panic!("walk failed with thread_count {thread_count:?}: {e}"));
-            parallel.sort_by(|a, b| a.path.cmp(&b.path));
-            assert_eq!(
-                parallel, sequential,
-                "thread_count {thread_count:?} produced a different file list after canonical sort"
-            );
-        }
-    }
-
-    /// A missing root is not an error — the bundle may legitimately not
-    /// exist yet (e.g. zccache restore produced nothing). The walk must
-    /// return an empty vec rather than propagating a `NotFound` IO error.
-    #[test]
-    fn walk_bundle_files_missing_root_returns_empty() {
-        let bundle = tempfile::tempdir().expect("tempdir");
-        let missing = bundle.path().join("never-created");
-        for thread_count in [Some(1), None, Some(4)] {
-            let entries = walk_bundle_files(&missing, thread_count)
-                .unwrap_or_else(|e| panic!("missing root must not error ({thread_count:?}): {e}"));
-            assert!(
-                entries.is_empty(),
-                "missing root walk with {thread_count:?} should be empty, got {entries:?}"
-            );
-        }
-    }
-}
+mod tests;
